==> 3_data_prep_analyses_20240201.Rmd <==
---
title: Genetic differentiation on flowering time in Cerastium fontanum using a greenhouse
  experiment
subtitle: Analyses suggested by Johan
author: "Alicia Valdés"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output:
  html_document:
    toc: yes
    toc_depth: '4'
    df_print: paged
  pdf_document:
    toc: yes
    toc_depth: 4
  word_document:
    toc: yes
    toc_depth: '4'
subtitle: Data preparation and first analyses
editor_options:
  chunk_output_type: inline
---

```{r setup, include=FALSE}
options(tibble.width = Inf)
knitr::opts_chunk$set(warning = FALSE, message = FALSE) 
```

```{r load packages, include=FALSE}
library(tidyverse)
library(readxl)
library(RColorBrewer)
library(ggeffects)
library(glmmTMB)
library(car)
library(sjPlot)
library(jtools)
library(broom)
library(broom.mixed)
library(knitr)
library(lme4)
library(lmerTest)
library(gridExtra)
library(ggthemes)
library(kableExtra)
library(ggpubr)
library(brms)
library(nadiv)
library(bayesplot)
library(matrixStats)
library(tidybayes)
library(cowplot)
library(performance)
library(partR2)
library(ggh4x)
```

```{r Define ggplot themes and palettes, include=FALSE}
my_theme <- function(){
  theme_base()+theme(plot.background=element_rect(fill="white", colour=NA))+
  theme(legend.position="none")+theme(text=element_text(family="serif"))+
  theme(plot.title = element_text(hjust =-0.06))
}
my_theme_legend <- function(){
  theme_base()+theme(plot.background=element_rect(fill="white", colour=NA))+
  theme(text=element_text(family="serif"))+
  theme(plot.title = element_text(hjust =-0.06))
}
myPalette <- colorRampPalette(brewer.pal(11, "YlOrRd"))
```

# Data preparation

## Read data from Excel files

```{r}
data_exp<-read_excel("data/edited/Cerastium_greenhouse_spring_2022_editedAV.xlsx", 
                       sheet="extracted_data")
data_exp$treat<-as.factor(data_exp$treat)
data_exp$dam<-as.factor(data_exp$mother)
data_exp$sire<-as.factor(data_exp$father)
data_exp$total_n_fl<-as.integer(data_exp$total_n_fl)
data_exp$date50<-as.numeric(data_exp$date50)
data_exp$unique_id<-as.factor(data_exp$unique_id)
data_exp<-data_exp%>%filter(unique_id!="122/144_41") 
# Remove one case where temperature of the parents is not known
```

```{r}
data_selfed<-read_excel("data/edited/Cerastium_greenhouse_spring_2022_editedAV.xlsx", 
                       sheet="extracted_data_parents")
data_selfed$treat<-as.factor(data_selfed$treat)
data_selfed$id_parent<-as.factor(data_selfed$id_parent)
data_selfed<-data_selfed%>%
  mutate(unique_id=paste("selfed",row_number(),sep="_"))
```

```{r}
data_shoots<-read_excel("data/edited/Cerastium_greenhouse_spring_2022_editedAV.xlsx", 
                       sheet="extracted_data_shoots")
data_shoots$unique_id<-as.factor(data_shoots$unique_id)
# Dates say 2022 and that is not true, but should be OK if we need to
# calculate differences
```

```{r}
blocks_dams<-read_excel("data/edited/crossing_design.xlsx", 
                       sheet="blocks_mothers")
blocks_sires<-read_excel("data/edited/crossing_design.xlsx", 
                       sheet="blocks_fathers")
```


```{r}
head(data_exp)
head(data_selfed)
head(data_shoots)
head(blocks_dams)
head(blocks_sires)
```

## Unique id is unique in all cases --> Fixed

```{r}
data_exp%>%group_by(unique_id)%>%summarise(count=n())%>%filter(count>1)
data_shoots%>%group_by(unique_id)%>%summarise(count=n())%>%filter(count>1)
```

## How many different dams and sires?

```{r}
length(levels(data_exp$dam)) # 25 different dams (shouldn't it be 24?)
length(levels(data_exp$sire)) # 24 different dams
```

## How many unique combinations of dams and sires?

```{r}
data_exp<-data_exp%>%mutate(dam_sire=paste(dam,sire,sep="_"))
```

```{r}
length(unique(data_exp$dam_sire)) # 99 (shouldn't it be 96?)
```

## Number of plants from each crossing

```{r}
data_exp%>%group_by(dam_sire)%>%summarise(num=n())%>%
  ggplot(aes(x=num))+geom_histogram(bins=12)
```

12 plants from most crossings, but also one with 13 and some with fewer.

## Adding shoot info

```{r}
data_shoots<-data_shoots%>%
  mutate(day_diff=date2-date1,
         mean_sh_h_date1=(sh1_date1+sh2_date1+sh3_date1)/3,
         mean_sh_h_date2=(sh1_date2+sh2_date2+sh3_date2)/3,
         mean_sh_growth=mean_sh_h_date2-mean_sh_h_date1)
```

```{r}
data_exp<-data_exp%>%left_join(data_shoots%>%select(unique_id,mean_sh_growth,
                                                    mean_sh_h_date1,
                                                    mean_sh_h_date2))
```

## Adding block info to crossings

```{r}
data_exp_blocks<-data_exp%>%
  left_join(blocks_dams%>%mutate(dam=factor(mother)))%>%
  left_join(blocks_sires%>%mutate(sire=factor(father)))%>%
  # Removing 2 cases where block_dam different from block_sire
  filter(block_mother==block_father)%>%
  mutate(block=factor(block_mother))%>%
  select(!(block_mother:block_father))
```

## Add block info to selfed

```{r}
data_selfed_blocks<-data_selfed%>%
  mutate(dam=id_parent,sire=id_parent)%>%
  select(-id_parent)%>%
  left_join(blocks_dams%>%mutate(dam=factor(mother)))%>%
  left_join(blocks_sires%>%mutate(sire=factor(father)))%>%
  mutate(block=factor(ifelse(is.na(block_mother),block_father,block_mother)))%>%
  select(!(block_mother:block_father))
```

# Change date 50 to MFD

```{r}
data_exp<-data_exp%>%rename(MFD=date50)
```

# Distributions

```{r}
hist(data_exp$FFD)
hist(data_exp$LFD)
hist(data_exp$MFD)
data_exp$duration<-data_exp$LFD-data_exp$FFD
hist(data_exp$duration)
hist(data_exp$mean_sh_growth)
```

All looking quite normal.

# Analyses

## Prediction 1

### Models (without flower number)

```{r}
model1a_veg_uh<-lmer(mean_sh_growth~1+(1|sire)+(1|dam),
                       #(1|sire:dam) # Removed to avoid singular fit
                     subset(data_exp,treat=="unheated"))
model1a_veg_h<-lmer(mean_sh_growth~1+(1|sire)+(1|dam)+(1|sire:dam),
                     subset(data_exp,treat=="heatmat"))
model1a_FFD_uh<-lmer(FFD~1+(1|sire)+(1|dam)+(1|sire:dam),
                 subset(data_exp,treat=="unheated"))
model1a_FFD_h<-lmer(FFD~1+(1|sire)+(1|dam)+(1|sire:dam),
                 subset(data_exp,treat=="heatmat"))
model1a_LFD_uh<-lmer(LFD~1+(1|sire)+(1|dam)+(1|sire:dam),
                 subset(data_exp,treat=="unheated"))
model1a_LFD_h<-lmer(LFD~1+(1|sire)+(1|dam)+(1|sire:dam),
                 subset(data_exp,treat=="heatmat"))
model1a_MFD_uh<-lmer(MFD~1+(1|sire)+(1|dam)+(1|sire:dam),
                 subset(data_exp,treat=="unheated"))
model1a_MFD_h<-lmer(MFD~1+(1|sire)+(1|dam)+(1|sire:dam),
                 subset(data_exp,treat=="heatmat"))
summary(model1a_veg_uh)
summary(model1a_veg_h)
summary(model1a_FFD_uh)
summary(model1a_FFD_h)
summary(model1a_LFD_uh)
summary(model1a_LFD_h)
summary(model1a_MFD_uh)
summary(model1a_MFD_h)
```

#### Proportions of variance

```{r}
Variance_veg_uh<-as.data.frame(VarCorr(model1a_veg_uh))[,c(1,4)]
Variance_FFD_uh<-as.data.frame(VarCorr(model1a_FFD_uh))[,c(1,4)]
Variance_LFD_uh<-as.data.frame(VarCorr(model1a_LFD_uh))[,c(1,4)]
Variance_MFD_uh<-as.data.frame(VarCorr(model1a_MFD_uh))[,c(1,4)]
Variance_veg_h<-as.data.frame(VarCorr(model1a_veg_h))[,c(1,4)]
Variance_FFD_h<-as.data.frame(VarCorr(model1a_FFD_h))[,c(1,4)]
Variance_LFD_h<-as.data.frame(VarCorr(model1a_LFD_h))[,c(1,4)]
Variance_MFD_h<-as.data.frame(VarCorr(model1a_MFD_h))[,c(1,4)]
# Intra-class correlations
PropVar_veg_uh <- Variance_veg_uh%>%mutate(propvar=vcov/sum(vcov))%>%
  mutate(variable="veg",treat="unheated")
PropVar_FFD_uh <- Variance_FFD_uh%>%mutate(propvar=vcov/sum(vcov))%>%
  mutate(variable="FFD",treat="unheated")
PropVar_LFD_uh <- Variance_LFD_uh%>%mutate(propvar=vcov/sum(vcov))%>%
  mutate(variable="LFD",treat="unheated")
PropVar_MFD_uh <- Variance_MFD_uh%>%mutate(propvar=vcov/sum(vcov))%>%
  mutate(variable="MFD",treat="unheated")
PropVar_veg_h <- Variance_veg_h%>%mutate(propvar=vcov/sum(vcov))%>%
  mutate(variable="veg",treat="heated")
PropVar_FFD_h <- Variance_FFD_h%>%mutate(propvar=vcov/sum(vcov))%>%
  mutate(variable="FFD",treat="heated")
PropVar_LFD_h <- Variance_LFD_h%>%mutate(propvar=vcov/sum(vcov))%>%
  mutate(variable="LFD",treat="heated")
PropVar_MFD_h <- Variance_MFD_h%>%mutate(propvar=vcov/sum(vcov))%>%
  mutate(variable="MFD",treat="heated")
Props_var<-rbind(PropVar_veg_uh,PropVar_FFD_uh,PropVar_LFD_uh,PropVar_MFD_uh,
                 PropVar_veg_h,PropVar_FFD_h,PropVar_LFD_h,PropVar_MFD_h)
```

```{r}
ggplot(Props_var,aes(x=variable,y=propvar,fill=grp))+
  geom_col()+facet_wrap(~treat)
```

#### Heritability and maternal effects

```{r}
# h^2 (paternal effects)
her_veg_uh<-as.numeric(4*subset(PropVar_veg_uh,grp=="sire")[3])
her_veg_h<-as.numeric(4*subset(PropVar_veg_h,grp=="sire")[3])
her_FFD_uh<-as.numeric(4*subset(PropVar_FFD_uh,grp=="sire")[3])
her_FFD_h<-as.numeric(4*subset(PropVar_FFD_h,grp=="sire")[3])
her_LFD_uh<-as.numeric(4*subset(PropVar_LFD_uh,grp=="sire")[3])
her_LFD_h<-as.numeric(4*subset(PropVar_LFD_h,grp=="sire")[3])
her_MFD_uh<-as.numeric(4*subset(PropVar_MFD_uh,grp=="sire")[3])
her_MFD_h<-as.numeric(4*subset(PropVar_MFD_h,grp=="sire")[3])
# Because the additive genetic variance, VA,
# is expected to be four times the among pollen‐donor variance 
# (Falconer & Mackay, 1996; Lynch & Walsh, 1998)
her<-data.frame(value=rbind(her_veg_uh,her_veg_h,her_FFD_uh,her_FFD_h,her_LFD_uh,
                       her_LFD_h,her_MFD_uh,her_MFD_h))%>%
  rownames_to_column()%>%
  mutate(variable=c("veg","veg","FFD","FFD","LFD","LFD","MFD","MFD"),
         treat=c("unheated","heated","unheated","heated","unheated","heated",
                 "unheated","heated"),
         effect="Heritability")
```

```{r}
# maternal effects

# Maternal - paternal
# ----------------------
# summed effects

# Because the pollen‐recipient variance component contains a combination of
# genetic and environmental effects, we subtracted the additive genetic 
# (pollen donor) component from the pollen-recipient variance component
# before dividing the resulting estimate by VP to estimate 
# m2 (m2 = (Vpollen recipient − Vpollen donor)/VP).

# But what about the interaction?

mat_veg_uh<-(Variance_veg_uh[1,2]-Variance_veg_uh[2,2])/
  as.numeric(Variance_veg_uh%>%summarise(sum(vcov)))
mat_veg_h<-(Variance_veg_h[2,2]-Variance_veg_h[3,2])/
  as.numeric(Variance_veg_h%>%summarise(sum(vcov)))
mat_FFD_uh<-(Variance_FFD_uh[2,2]-Variance_FFD_uh[3,2])/
  as.numeric(Variance_FFD_uh%>%summarise(sum(vcov)))
mat_FFD_h<-(Variance_FFD_h[2,2]-Variance_FFD_h[3,2])/
  as.numeric(Variance_FFD_h%>%summarise(sum(vcov)))
mat_LFD_uh<-(Variance_LFD_uh[2,2]-Variance_LFD_uh[3,2])/
  as.numeric(Variance_LFD_uh%>%summarise(sum(vcov)))
mat_LFD_h<-(Variance_LFD_h[2,2]-Variance_LFD_h[3,2])/
  as.numeric(Variance_LFD_h%>%summarise(sum(vcov)))
mat_MFD_uh<-(Variance_MFD_uh[2,2]-Variance_MFD_uh[3,2])/
  as.numeric(Variance_MFD_uh%>%summarise(sum(vcov)))
mat_MFD_h<-(Variance_MFD_h[2,2]-Variance_MFD_h[3,2])/
  as.numeric(Variance_MFD_h%>%summarise(sum(vcov)))
mat<-data.frame(value=rbind(mat_veg_uh,mat_veg_h,mat_FFD_uh,mat_FFD_h,mat_LFD_uh,
                       mat_LFD_h,mat_MFD_uh,mat_MFD_h))%>%
  rownames_to_column()%>%
  mutate(variable=c("veg","veg","FFD","FFD","LFD","LFD","MFD","MFD"),
         treat=c("unheated","heated","unheated","heated","unheated","heated",
                 "unheated","heated"),
         effect="Maternal effects")
```

```{r}
her_mat<-rbind(her,mat)
her_mat
```

```{r}
ggplot(her_mat,aes(x=variable,y=value,fill=effect))+
  geom_bar(stat="identity",position="dodge")+facet_wrap(~treat)
```

#### Appendix S3: LRTs for variance components

(These results could be used to add asterisks to the previous plot).

```{r}
ranova(model1a_veg_uh)
ranova(model1a_veg_h)
ranova(model1a_FFD_uh)
ranova(model1a_FFD_h)
ranova(model1a_LFD_uh)
ranova(model1a_LFD_h)
ranova(model1a_MFD_uh)
ranova(model1a_MFD_h)
```

### Models (with flower number)

```{r}
model1a_veg_uh_fln<-lmer(mean_sh_growth~total_n_fl+(1|sire)+(1|dam),
                       #(1|sire:dam) # Removed to avoid singular fit
                     subset(data_exp,treat=="unheated"))
model1a_veg_h_fln<-lmer(mean_sh_growth~total_n_fl+(1|sire)+(1|dam)+(1|sire:dam),
                     subset(data_exp,treat=="heatmat"))
model1a_FFD_uh_fln<-lmer(FFD~total_n_fl+(1|sire)+(1|dam)+(1|sire:dam),
                 subset(data_exp,treat=="unheated"))
model1a_FFD_h_fln<-lmer(FFD~total_n_fl+(1|sire)+(1|dam)+(1|sire:dam),
                 subset(data_exp,treat=="heatmat"))
model1a_LFD_uh_fln<-lmer(LFD~total_n_fl+(1|sire)+(1|dam)+(1|sire:dam),
                 subset(data_exp,treat=="unheated"))
model1a_LFD_h_fln<-lmer(LFD~total_n_fl+(1|dam)+(1|sire:dam),
                        #(1|sire) # Removed to avoid singular fit
                 subset(data_exp,treat=="heatmat"))
model1a_MFD_uh_fln<-lmer(MFD~total_n_fl+(1|sire)+(1|dam)+(1|sire:dam),
                 subset(data_exp,treat=="unheated"))
model1a_MFD_h_fln<-lmer(MFD~total_n_fl+(1|sire)+(1|dam)+(1|sire:dam),
                 subset(data_exp,treat=="heatmat"))
summary(model1a_veg_uh_fln)
summary(model1a_veg_h_fln)
summary(model1a_FFD_uh_fln)
summary(model1a_FFD_h_fln)
summary(model1a_LFD_uh_fln)
summary(model1a_LFD_h_fln)
summary(model1a_MFD_uh_fln)
summary(model1a_MFD_h_fln)
```

#### Proportions of variance

```{r}
Variance_veg_uh_fln<-as.data.frame(VarCorr(model1a_veg_uh_fln))[,c(1,4)]
Variance_FFD_uh_fln<-as.data.frame(VarCorr(model1a_FFD_uh_fln))[,c(1,4)]
Variance_LFD_uh_fln<-as.data.frame(VarCorr(model1a_LFD_uh_fln))[,c(1,4)]
Variance_MFD_uh_fln<-as.data.frame(VarCorr(model1a_MFD_uh_fln))[,c(1,4)]
Variance_veg_h_fln<-as.data.frame(VarCorr(model1a_veg_h_fln))[,c(1,4)]
Variance_FFD_h_fln<-as.data.frame(VarCorr(model1a_FFD_h_fln))[,c(1,4)]
Variance_LFD_h_fln<-as.data.frame(VarCorr(model1a_LFD_h_fln))[,c(1,4)]
Variance_MFD_h_fln<-as.data.frame(VarCorr(model1a_MFD_h_fln))[,c(1,4)]
# Intra-class correlations
PropVar_veg_uh_fln <- Variance_veg_uh_fln%>%mutate(propvar=vcov/sum(vcov))%>%
  mutate(variable="veg",treat="unheated")
PropVar_FFD_uh_fln <- Variance_FFD_uh_fln%>%mutate(propvar=vcov/sum(vcov))%>%
  mutate(variable="FFD",treat="unheated")
PropVar_LFD_uh_fln <- Variance_LFD_uh_fln%>%mutate(propvar=vcov/sum(vcov))%>%
  mutate(variable="LFD",treat="unheated")
PropVar_MFD_uh_fln <- Variance_MFD_uh_fln%>%mutate(propvar=vcov/sum(vcov))%>%
  mutate(variable="MFD",treat="unheated")
PropVar_veg_h_fln <- Variance_veg_h_fln%>%mutate(propvar=vcov/sum(vcov))%>%
  mutate(variable="veg",treat="heated")
PropVar_FFD_h_fln <- Variance_FFD_h_fln%>%mutate(propvar=vcov/sum(vcov))%>%
  mutate(variable="FFD",treat="heated")
PropVar_LFD_h_fln <- Variance_LFD_h_fln%>%mutate(propvar=vcov/sum(vcov))%>%
  mutate(variable="LFD",treat="heated")
PropVar_MFD_h_fln <- Variance_MFD_h_fln%>%mutate(propvar=vcov/sum(vcov))%>%
  mutate(variable="MFD",treat="heated")
Props_var_fln<-rbind(PropVar_veg_uh_fln,PropVar_FFD_uh_fln,PropVar_LFD_uh_fln,
                     PropVar_MFD_uh_fln,PropVar_veg_h_fln,PropVar_FFD_h_fln,
                     PropVar_LFD_h_fln,PropVar_MFD_h_fln)
```

```{r}
ggplot(Props_var_fln,aes(x=variable,y=propvar,fill=grp))+
  geom_col()+facet_wrap(~treat)
```

#### Heritability and maternal effects

```{r}
# h^2 (paternal effects)
her_veg_uh_fln<-as.numeric(4*subset(PropVar_veg_uh_fln,grp=="sire")[3])
her_veg_h_fln<-as.numeric(4*subset(PropVar_veg_h_fln,grp=="sire")[3])
her_FFD_uh_fln<-as.numeric(4*subset(PropVar_FFD_uh_fln,grp=="sire")[3])
her_FFD_h_fln<-as.numeric(4*subset(PropVar_FFD_h_fln,grp=="sire")[3])
her_LFD_uh_fln<-as.numeric(4*subset(PropVar_LFD_uh_fln,grp=="sire")[3])
her_LFD_h_fln<-as.numeric(4*subset(PropVar_LFD_h_fln,grp=="sire")[3])
her_MFD_uh_fln<-as.numeric(4*subset(PropVar_MFD_uh_fln,grp=="sire")[3])
her_MFD_h_fln<-as.numeric(4*subset(PropVar_MFD_h_fln,grp=="sire")[3])
# Because the additive genetic variance, VA,
# is expected to be four times the among pollen‐donor variance 
# (Falconer & Mackay, 1996; Lynch & Walsh, 1998)
her_fln<-data.frame(value=rbind(her_veg_uh_fln,her_veg_h_fln,her_FFD_uh_fln,
                                her_FFD_h_fln,her_LFD_uh_fln,her_LFD_h_fln,
                                her_MFD_uh_fln,her_MFD_h_fln))%>%
  rownames_to_column()%>%
  mutate(variable=c("veg","veg","FFD","FFD","LFD","LFD","MFD","MFD"),
         treat=c("unheated","heated","unheated","heated","unheated","heated",
                 "unheated","heated"),
         effect="Heritability")
```

```{r}
# maternal effects

# Maternal - paternal
# ----------------------
# summed effects

# Because the pollen‐recipient variance component contains a combination of
# genetic and environmental effects, we subtracted the additive genetic 
# (pollen donor) component from the pollen-recipient variance component
# before dividing the resulting estimate by VP to estimate 
# m2 (m2 = (Vpollen recipient − Vpollen donor)/VP).

# But what about the interaction?

mat_veg_uh_fln<-(Variance_veg_uh_fln[1,2]-Variance_veg_uh_fln[2,2])/
  as.numeric(Variance_veg_uh_fln%>%summarise(sum(vcov)))
mat_veg_h_fln<-(Variance_veg_h_fln[2,2]-Variance_veg_h_fln[3,2])/
  as.numeric(Variance_veg_h_fln%>%summarise(sum(vcov)))
mat_FFD_uh_fln<-(Variance_FFD_uh_fln[2,2]-Variance_FFD_uh_fln[3,2])/
  as.numeric(Variance_FFD_uh_fln%>%summarise(sum(vcov)))
mat_FFD_h_fln<-(Variance_FFD_h_fln[2,2]-Variance_FFD_h_fln[3,2])/
  as.numeric(Variance_FFD_h_fln%>%summarise(sum(vcov)))
mat_LFD_uh_fln<-(Variance_LFD_uh_fln[2,2]-Variance_LFD_uh_fln[3,2])/
  as.numeric(Variance_LFD_uh_fln%>%summarise(sum(vcov)))
mat_LFD_h_fln<-(Variance_LFD_h_fln[2,2]-Variance_LFD_h_fln[3,2])/
  as.numeric(Variance_LFD_h_fln%>%summarise(sum(vcov)))
mat_MFD_uh_fln<-(Variance_MFD_uh_fln[2,2]-Variance_MFD_uh_fln[3,2])/
  as.numeric(Variance_MFD_uh_fln%>%summarise(sum(vcov)))
mat_MFD_h_fln<-(Variance_MFD_h_fln[2,2]-Variance_MFD_h_fln[3,2])/
  as.numeric(Variance_MFD_h_fln%>%summarise(sum(vcov)))
mat_fln<-data.frame(value=rbind(mat_veg_uh_fln,mat_veg_h_fln,mat_FFD_uh_fln,
                            mat_FFD_h_fln,mat_LFD_uh_fln,mat_LFD_h_fln,
                            mat_MFD_uh_fln,mat_MFD_h_fln))%>%
  rownames_to_column()%>%
  mutate(variable=c("veg","veg","FFD","FFD","LFD","LFD","MFD","MFD"),
         treat=c("unheated","heated","unheated","heated","unheated","heated",
                 "unheated","heated"),
         effect="Maternal effects")
```

```{r}
her_mat_fln<-rbind(her_fln,mat_fln)
her_mat_fln
```

```{r}
ggplot(her_mat_fln,aes(x=variable,y=value,fill=effect))+
  geom_bar(stat="identity",position="dodge")+facet_wrap(~treat)
```

```{r}
ggplot(rbind(her_mat%>%mutate(fln="Without flower nr"),
             her_mat_fln%>%mutate(fln="With flower nr")),
       aes(x=variable,y=value,fill=fln))+
  geom_bar(stat="identity",position="dodge")+
  facet_wrap(~treat+effect)
```

## Prediction 2

### Model (without flower number)

```{r}
model1b_veg<-lmer(mean_sh_growth~treat+
                    (1|sire)+(1|dam)+
                    # We still need sire and dam, right?
                    #(1|treat:sire)+ # Removed to avoid singular fit
                    (1|treat:dam)+
                    (1|sire:dam),
                  data_exp)
model1b_FFD<-lmer(FFD~treat+
                    (1|sire)+(1|dam)+
                    # We still need sire and dam, right?
                    (1|treat:sire)+
                    #(1|treat:dam)+ # Removed to avoid singular fit
                    (1|sire:dam),
                  data_exp)
model1b_LFD<-lmer(LFD~treat+
                    (1|sire)+(1|dam)+
                    # We still need sire and dam, right?
                    #(1|treat:sire)+ # Removed to avoid singular fit
                    (1|treat:dam)+
                    (1|sire:dam),
                  data_exp)
model1b_MFD<-lmer(MFD~treat+
                    (1|sire)+(1|dam)+
                    # We still need sire and dam, right?
                    (1|treat:sire)+
                    #(1|treat:dam)+ 2.330e-07
                    (1|sire:dam),
                  data_exp)
summary(model1b_veg)
summary(model1b_FFD)
summary(model1b_LFD)
summary(model1b_MFD)
```

#### Proportions of variance

```{r}
# extract variance components
Variance_veg_treat <- as.data.frame(VarCorr(model1b_veg))[,c(1,4)]  
Variance_FFD_treat <- as.data.frame(VarCorr(model1b_FFD))[,c(1,4)]  
Variance_LFD_treat <- as.data.frame(VarCorr(model1b_LFD))[,c(1,4)]  
Variance_MFD_treat <- as.data.frame(VarCorr(model1b_MFD))[,c(1,4)]  
# Intra-class correlation
PropVar_veg_treat <- Variance_veg_treat%>%
  mutate(propvar=vcov/sum(vcov))%>%mutate(variable="veg")
PropVar_FFD_treat <- Variance_FFD_treat%>%
  mutate(propvar=vcov/sum(vcov))%>%mutate(variable="FFD")
PropVar_LFD_treat <- Variance_LFD_treat%>%
  mutate(propvar=vcov/sum(vcov))%>%mutate(variable="LFD")
PropVar_MFD_treat <- Variance_MFD_treat%>%
  mutate(propvar=vcov/sum(vcov))%>%mutate(variable="MFD")
# Proportional variance
Props_var_treat<-rbind(PropVar_veg_treat,PropVar_FFD_treat,PropVar_LFD_treat,
                       PropVar_MFD_treat)
```

```{r}
ggplot(Props_var_treat,aes(x=variable,y=propvar,fill=grp))+
  geom_col()
```

#### Heritability and maternal effects

Are these heritability and maternal effects of the SLOPE of RNs?

```{r}
# h^2 (paternal effects)
her_veg_treat<-0
her_FFD_treat<-4*subset(PropVar_FFD_treat,grp=="treat:sire")[3]
her_LFD_treat<-0
her_MFD_treat<-4*subset(PropVar_MFD_treat,grp=="treat:sire")[3]
# Because the additive genetic variance, VA,
# is expected to be four times the among pollen‐donor variance 
# (Falconer & Mackay, 1996; Lynch & Walsh, 1998)
her_treat<-data.frame(value=rbind(her_veg_treat,her_FFD_treat,
                                  her_LFD_treat,her_MFD_treat))%>%
  mutate(variable=c("veg","FFD","LFD","MFD"),
         effect="Heritability")%>%
  rename(value=propvar)
```

```{r}
# maternal effects

# Maternal - paternal
# ----------------------
# summed effects

# Because the pollen‐recipient variance component contains a combination of
# genetic and environmental effects, we subtracted the additive genetic 
# (pollen donor) component from the pollen-recipient variance component
# before dividing the resulting estimate by VP to estimate 
# m2 (m2 = (Vpollen recipient − Vpollen donor)/VP).

mat_veg_treat<-(subset(Variance_veg_treat,grp=="treat:dam")[2]-
                  0)/
  Variance_veg_treat%>%summarise(sum(vcov))
mat_FFD_treat<-(0-
                  subset(Variance_FFD_treat,grp=="treat:sire")[2])/
  Variance_FFD_treat%>%summarise(sum(vcov))
mat_LFD_treat<-(subset(Variance_LFD_treat,grp=="treat:dam")[2]-
                  0)/
  Variance_LFD_treat%>%summarise(sum(vcov))
mat_MFD_treat<-(0-
                  subset(Variance_MFD_treat,grp=="treat:sire")[2])/
  Variance_MFD_treat%>%summarise(sum(vcov))
mat_treat<-data.frame(value=rbind(mat_veg_treat,mat_FFD_treat,
                                  mat_LFD_treat,mat_MFD_treat))%>%
  mutate(variable=c("veg","FFD","LFD","MFD"),
         effect="Maternal effects")%>%
  rename(value=vcov)
```

```{r}
her_mat_treat<-rbind(her_treat,mat_treat)
her_mat_treat
```

```{r}
ggplot(her_mat_treat,aes(x=variable,y=value,fill=effect))+
  geom_bar(stat="identity",position="dodge")
```

This is probably not correct! Can we really get negative maternal effects? YES?

#### Appendix S4: LRTs for variance components

(These results could be used to add asterisks to the previous plot).

```{r}
ranova(model1b_veg)
ranova(model1b_FFD)
ranova(model1b_LFD)
ranova(model1b_MFD)
```

### Model (with flower number)

```{r}
model1b_veg_fln<-lmer(mean_sh_growth~treat+total_n_fl+
                    (1|sire)+(1|dam)+
                    # We still need sire and dam, right?
                    #(1|treat:sire)+ # Removed to avoid singular fit
                    (1|treat:dam)+
                    (1|sire:dam),
                  data_exp)
model1b_FFD_fln<-lmer(FFD~treat+total_n_fl+
                    (1|sire)+(1|dam)+
                    # We still need sire and dam, right?
                    (1|treat:sire)+
                    #(1|treat:dam)+ # Removed to avoid singular fit
                    (1|sire:dam),
                  data_exp)
model1b_LFD_fln<-lmer(LFD~treat+total_n_fl+
                    (1|sire)+(1|dam)+
                    # We still need sire and dam, right?
                    #(1|treat:sire)+ # Removed to avoid singular fit
                    (1|treat:dam)+
                    (1|sire:dam),
                  data_exp)
model1b_MFD_fln<-lmer(MFD~treat+total_n_fl+
                    (1|sire)+(1|dam)+
                    # We still need sire and dam, right?
                    (1|treat:sire)+(1|treat:dam)+
                    (1|sire:dam),
                  data_exp)
summary(model1b_veg_fln)
summary(model1b_FFD_fln)
summary(model1b_LFD_fln)
summary(model1b_MFD_fln)
```

#### Proportions of variance

```{r}
# extract variance components
Variance_veg_treat_fln <- as.data.frame(VarCorr(model1b_veg_fln))[,c(1,4)]  
Variance_FFD_treat_fln <- as.data.frame(VarCorr(model1b_FFD_fln))[,c(1,4)]  
Variance_LFD_treat_fln <- as.data.frame(VarCorr(model1b_LFD_fln))[,c(1,4)]  
Variance_MFD_treat_fln <- as.data.frame(VarCorr(model1b_MFD_fln))[,c(1,4)]  
# Intra-class correlation
PropVar_veg_treat_fln <- Variance_veg_treat_fln%>%
  mutate(propvar=vcov/sum(vcov))%>%mutate(variable="veg")
PropVar_FFD_treat_fln <- Variance_FFD_treat_fln%>%
  mutate(propvar=vcov/sum(vcov))%>%mutate(variable="FFD")
PropVar_LFD_treat_fln <- Variance_LFD_treat_fln%>%
  mutate(propvar=vcov/sum(vcov))%>%mutate(variable="LFD")
PropVar_MFD_treat_fln <- Variance_MFD_treat_fln%>%
  mutate(propvar=vcov/sum(vcov))%>%mutate(variable="MFD")
# Proportional variance
Props_var_treat_fln<-rbind(PropVar_veg_treat_fln,PropVar_FFD_treat_fln,
                           PropVar_LFD_treat_fln,PropVar_MFD_treat_fln)
```

```{r}
ggplot(Props_var_treat_fln,aes(x=variable,y=propvar,fill=grp))+
  geom_col()
```

#### Heritability and maternal effects

Are these heritability and maternal effects of the SLOPE of RNs?

```{r}
# h^2 (paternal effects)
her_veg_treat_fln<-0
her_FFD_treat_fln<-4*subset(PropVar_FFD_treat_fln,grp=="treat:sire")[3]
her_LFD_treat_fln<-0
her_MFD_treat_fln<-4*subset(PropVar_MFD_treat_fln,grp=="treat:sire")[3]
# Because the additive genetic variance, VA,
# is expected to be four times the among pollen‐donor variance 
# (Falconer & Mackay, 1996; Lynch & Walsh, 1998)
her_treat_fln<-data.frame(value=rbind(her_veg_treat_fln,her_FFD_treat_fln,
                                  her_LFD_treat_fln,her_MFD_treat_fln))%>%
  mutate(variable=c("veg","FFD","LFD","MFD"),
         effect="Heritability")%>%
  rename(value=propvar)
```

```{r}
# maternal effects

# Maternal - paternal
# ----------------------
# summed effects

# Because the pollen‐recipient variance component contains a combination of
# genetic and environmental effects, we subtracted the additive genetic 
# (pollen donor) component from the pollen-recipient variance component
# before dividing the resulting estimate by VP to estimate 
# m2 (m2 = (Vpollen recipient − Vpollen donor)/VP).

mat_veg_treat_fln<-(subset(Variance_veg_treat_fln,grp=="treat:dam")[2]-
                  0)/
  Variance_veg_treat_fln%>%summarise(sum(vcov))
mat_FFD_treat_fln<-(0-
                  subset(Variance_FFD_treat_fln,grp=="treat:sire")[2])/
  Variance_FFD_treat_fln%>%summarise(sum(vcov))
mat_LFD_treat_fln<-(subset(Variance_LFD_treat_fln,grp=="treat:dam")[2]-
                  0)/
  Variance_LFD_treat_fln%>%summarise(sum(vcov))
mat_MFD_treat_fln<-(0-
                  subset(Variance_MFD_treat_fln,grp=="treat:sire")[2])/
  Variance_MFD_treat_fln%>%summarise(sum(vcov))
mat_treat_fln<-data.frame(value=rbind(mat_veg_treat_fln,mat_FFD_treat_fln,
                                  mat_LFD_treat_fln,mat_MFD_treat_fln))%>%
  mutate(variable=c("veg","FFD","LFD","MFD"),
         effect="Maternal effects")%>%
  rename(value=vcov)
```

```{r}
her_mat_treat_fln<-rbind(her_treat_fln,mat_treat_fln)
her_mat_treat_fln
```

```{r}
ggplot(her_mat_treat_fln,aes(x=variable,y=value,fill=effect))+
  geom_bar(stat="identity",position="dodge")
```

```{r}
ggplot(rbind(her_mat_treat%>%mutate(fln="Without flower nr"),
             her_mat_treat_fln%>%mutate(fln="With flower nr")),
       aes(x=variable,y=value,fill=fln))+
  geom_bar(stat="identity",position="dodge")+
  facet_wrap(~effect)
```

## Figure 1 (Predictions 1-2)

```{r}
Props_var<-Props_var%>%
  mutate(variable=factor(variable,levels=c("veg","FFD","MFD","LFD")),
         treat=factor(treat,levels=c("unheated","heated")),
         grp=factor(grp,levels=c("sire","dam","sire:dam","Residual")),
         sig=c("y","y","y","n","y","y","n",
               "n","y","n","n","n","y","y","n",
               "n","y","n","n","y","y","n","n",
               "n","y","n","n","n","y","n","n"))%>% # Based on LRTs
  mutate(treat=fct_recode(treat,"control"="unheated"))
Props_var_treat<-Props_var_treat%>%
  mutate(variable=factor(variable,levels=c("veg","FFD","MFD","LFD")),
         grp=factor(grp,levels=c("sire","dam","sire:dam",
                                 "treat:sire","treat:dam","Residual")),
         sig=c("y","n","y","y","n",
               "n","y","y","n","n",
               "n","n","y","n","n",
               "n","y","y","n","n")) # Based on LRTs
```

```{r}
legend_fig1<-get_legend(
  ggplot(Props_var_treat%>%filter(grp!="Residual"),
         aes(x=variable,y=propvar,fill=grp,shape=sig))+
    geom_col(color="black",size=0.3)+
    geom_point(aes(x=variable,y=propvar,group=grp),
               position=position_stack(vjust=.5),show.legend=F,size=2)+
    my_theme_legend()+
    xlab(NULL)+ylab("Proportion of variance")+
    theme(legend.position="top",legend.title=element_blank())+
    scale_fill_discrete(labels=c("sire","dam","sire:dam",
                                 "treat:sire","treat:dam","residual"),
                        type=c("#00BFC4","#F8766D","#00BA38","#619CFF",
                               "#F564E3","grey"))+
    scale_shape_manual(values=c(NA,8))+ggtitle("B)")+
    theme(axis.text.y=element_blank(),axis.title.y=element_blank()))
fig1a<-ggplot(Props_var%>%filter(grp!="Residual"),
              aes(x=variable,y=propvar,fill=grp,shape=sig))+
  geom_col(color="black",size=0.3)+
  geom_point(aes(x=variable,y=propvar,group=grp),
             position=position_stack(vjust=.5),show.legend=F,size=2)+
  facet_wrap(~treat)+
  my_theme()+xlab(NULL)+ylab("Proportion of variance")+
  scale_fill_discrete(labels=c("sire","dam","sire:dam","residual"),
                      type=c("#00BFC4","#F8766D","#00BA38","grey"))+
  guides(fill=guide_legend(nrow=2))+
  theme(plot.title = element_text(hjust =-0.05,vjust=0))+
  scale_shape_manual(values=c(NA,8))+ggtitle("A)")
fig1b<-ggplot(Props_var_treat%>%filter(grp!="Residual"),
              aes(x=variable,y=propvar,fill=grp,shape=sig))+
  geom_col(color="black",size=0.3)+
  geom_point(aes(x=variable,y=propvar,group=grp),
             position=position_stack(vjust=.5),show.legend=F,size=2)+
  my_theme()+
  xlab(NULL)+ylab("Proportion of variance")+
  scale_fill_discrete(labels=c("sire","dam","sire:dam",
                               "treat:sire","treat:dam","residual"),
                      type=c("#00BFC4","#F8766D","#00BA38","#619CFF",
                             "#F564E3","grey"))+
  theme(plot.title = element_text(hjust =-0.05,vjust=4))+
  scale_shape_manual(values=c(NA,8))+ggtitle("B)")+
  theme(axis.title.y=element_blank())
fig1_without_legend<-cowplot::plot_grid(fig1a,fig1b,
                                        ncol=2,align="hv",axis="t",
                                        rel_widths = c(1.7, 1))
fig1<-grid.arrange(legend_fig1,fig1_without_legend,ncol=1,
                   heights=c(0.05,0.95))
plot(fig1)
ggsave(file="output/figures/fig1.tiff", plot=fig1,width=10,height=5,dpi=300,
       compression="lzw")
```

## Table 1 (Predictions 1-2)

```{r}
her_mat<-her_mat%>%
  mutate(variable=factor(variable,levels=c("veg","FFD","MFD","LFD")),
         treat=factor(treat,levels=c("unheated","heated")))%>%
  select(-rowname)%>%
  pivot_wider(names_from="effect",values_from="value")
her_mat
her_mat_treat<-her_mat_treat%>%
  pivot_wider(names_from="effect",values_from="value")
her_mat_treat
```

## Figure 2 (Prediction 2)

Main effect of temperature treatment NS for vegetative growth, and significant for FFD, LFD and MFD.

```{r}
data_fig2<-rbind(
  data.frame(ggpredict(model1b_veg,terms="treat"))%>%mutate(variable="veg"),
  data.frame(ggpredict(model1b_FFD,terms="treat"))%>%mutate(variable="FFD"),
  data.frame(ggpredict(model1b_MFD,terms="treat"))%>%mutate(variable="MFD"),
  data.frame(ggpredict(model1b_LFD,terms="treat"))%>%mutate(variable="LFD"),
  data.frame(ggpredict(model1b_FFD,terms=c("treat","sire"),type="random"))%>%
    mutate(variable="FFD"),
  data.frame(ggpredict(model1b_MFD,terms=c("treat","sire"),type="random"))%>%
    mutate(variable="MFD"))%>%
  mutate(variable=factor(variable,levels=c("veg","FFD","MFD","LFD")),
         x=factor(x,levels=c("unheated","heatmat"),
                  labels=c("unheated","heated")),
         type=ifelse(group==1,"mean","individual"))
```

```{r}
# fig2<-ggplot()+
#   geom_point(data=data_fig2%>%filter(group==1)%>%
#                mutate(x=factor(x,levels=c("unheated","heated")))%>%
#                mutate(x=fct_recode(x,"control"="unheated")),
#              aes(x=x,y=predicted),size=2)+
#   geom_errorbar(data=data_fig2%>%filter(group==1)%>%
#                   mutate(x=factor(x,levels=c("unheated","heated")))%>%
#                mutate(x=fct_recode(x,"control"="unheated")),
#                 aes(x=x,y=predicted,ymin=conf.low,ymax=conf.high),width=0.05)+
#   geom_line(data=data_fig2%>%filter(group==1)%>%
#               mutate(x=factor(x,levels=c("unheated","heated")))%>%
#                mutate(x=fct_recode(x,"control"="unheated")),
#             aes(x=x,y=predicted,group=variable))+
#   geom_line(data=data_fig2%>%filter(group!=1)%>%
#               mutate(x=factor(x,levels=c("unheated","heated")))%>%
#                mutate(x=fct_recode(x,"control"="unheated")),
#             aes(x=x,y=predicted,group=group,color=group),linewidth=0.2)+
#   geom_point(data=data_fig2%>%filter(group==1)%>%
#                mutate(x=factor(x,levels=c("unheated","heated")))%>%
#                mutate(x=fct_recode(x,"control"="unheated")),
#              aes(x=x,y=predicted),size=3)+
#   geom_errorbar(data=data_fig2%>%filter(group==1)%>%
#                   mutate(x=factor(x,levels=c("unheated","heated")))%>%
#                mutate(x=fct_recode(x,"control"="unheated")),
#                 aes(x=x,y=predicted,ymin=conf.low,ymax=conf.high),
#                 width=0.05)+
#   geom_line(data=data_fig2%>%filter(group==1)%>%
#               mutate(x=factor(x,levels=c("unheated","heated")))%>%
#                mutate(x=fct_recode(x,"control"="unheated")),
#             aes(x=x,y=predicted,group=variable))+
#   facet_wrap(~variable,ncol=2,scales="free")+my_theme()+
#   xlab(NULL)+ylab("Value")+scale_x_discrete(expand=c(0.1,0.1))
# fig2
# ggsave(file="output/figures/fig2.tiff", plot=fig2,width=6,height=5,dpi=300,
#        compression="lzw")
```

```{r}
fig2<-grid.arrange(
  # vegetative growth
  ggplot()+
    geom_point(data=data_fig2%>%filter(group==1)%>%
                 mutate(x=factor(x,levels=c("unheated","heated")))%>%
                 mutate(x=fct_recode(x,"control"="unheated"))%>%
                 filter(variable=="veg"),
               aes(x=x,y=predicted),size=2)+
    geom_errorbar(data=data_fig2%>%filter(group==1)%>%
                    mutate(x=factor(x,levels=c("unheated","heated")))%>%
                    mutate(x=fct_recode(x,"control"="unheated"))%>%
                    filter(variable=="veg"),
                  aes(x=x,y=predicted,ymin=conf.low,ymax=conf.high),
                  width=0.05,size=0.5)+
    geom_line(data=data_fig2%>%filter(group==1)%>%
                mutate(x=factor(x,levels=c("unheated","heated")))%>%
                mutate(x=fct_recode(x,"control"="unheated"))%>%
                filter(variable=="veg"),
              aes(x=x,y=predicted,group=variable),size=0.5)+
    my_theme()+ggtitle("A)")+
    theme(plot.title = element_text(hjust =-0.2,vjust=-1.5))+
    xlab(NULL)+ylab("Vegetative growth")+scale_x_discrete(expand=c(0.1,0.1)),
  # flowering time
  ggplot()+
    # Mean responses
    geom_point(data=data_fig2%>%filter(group==1)%>%
                 mutate(x=factor(x,levels=c("unheated","heated")))%>%
                 mutate(x=fct_recode(x,"control"="unheated"))%>%
                 filter(variable!="veg"),
               aes(x=x,y=predicted,color=variable),size=2)+
    geom_errorbar(data=data_fig2%>%filter(group==1)%>%
                    mutate(x=factor(x,levels=c("unheated","heated")))%>%
                    mutate(x=fct_recode(x,"control"="unheated"))%>%
                    filter(variable!="veg"),
                  aes(x=x,y=predicted,ymin=conf.low,ymax=conf.high,color=variable),
                  width=0.05,size=0.5)+
    geom_line(data=data_fig2%>%filter(group==1)%>%
                mutate(x=factor(x,levels=c("unheated","heated")))%>%
                mutate(x=fct_recode(x,"control"="unheated"))%>%
                filter(variable!="veg"),
              aes(x=x,y=predicted,group=variable,color=variable),size=0.5)+
    # Response of each sire for FFD
    geom_line(data=data_fig2%>%filter(group!=1)%>%
                mutate(x=factor(x,levels=c("unheated","heated")))%>%
                mutate(x=fct_recode(x,"control"="unheated"))%>%
                filter(variable=="FFD"),
              aes(x=x,y=predicted,group=group),
              linewidth=0.2,color="#F8766D",alpha=0.5)+
    # Response of each sire for MFD
    geom_line(data=data_fig2%>%filter(group!=1)%>%
                mutate(x=factor(x,levels=c("unheated","heated")))%>%
                mutate(x=fct_recode(x,"control"="unheated"))%>%
                filter(variable=="MFD"),
              aes(x=x,y=predicted,group=group),
              linewidth=0.2,color="#00BA38",alpha=0.5)+
    my_theme_legend()+ggtitle("B)")+
    theme(legend.title=element_blank())+
    theme(plot.title = element_text(hjust =-0.2,vjust=-1.5))+
    theme(legend.position = c(0.84, 0.88))+
    theme(legend.background = element_rect(fill = "transparent"))+
    xlab(NULL)+ylab("Flowering date")+scale_x_discrete(expand=c(0.1,0.1)),
  ncol=2)
fig2
ggsave(file="output/figures/fig2.tiff", plot=fig2,width=8,height=4,dpi=300,
       compression="lzw")
```


## Prediction 3

### Models

```{r}
data_exp<-data_exp%>%
  rename(temp_sire=temp_father,temp_dam=temp_mother)
```

```{r}
model3_veg_uh<-lmer(mean_sh_growth~temp_sire+temp_dam+
                      (1|sire)+(1|dam),
                      #(1|sire:dam) # Removed to avoid singular fit
                 subset(data_exp,treat=="unheated"&!is.na(mean_sh_growth)))
model3_veg_h<-lmer(mean_sh_growth~temp_sire+temp_dam+
                     (1|sire)+(1|dam)+(1|sire:dam),
                 subset(data_exp,treat=="heatmat"&!is.na(mean_sh_growth)))
model3_FFD_uh<-lmer(FFD~temp_sire+temp_dam+
                      (1|sire)+(1|dam)+(1|sire:dam),
                 subset(data_exp,treat=="unheated"&!is.na(FFD)))
model3_FFD_h<-lmer(FFD~temp_sire+temp_dam+
                     (1|sire)+(1|dam)+(1|sire:dam),
                 subset(data_exp,treat=="heatmat"&!is.na(FFD)))
model3_LFD_uh<-lmer(LFD~temp_sire+temp_dam+
                      (1|sire)+(1|dam)+(1|sire:dam),
                 subset(data_exp,treat=="unheated"&!is.na(LFD)))
model3_LFD_h<-lmer(LFD~temp_sire+temp_dam+
                     (1|sire)+(1|dam)+(1|sire:dam),
                 subset(data_exp,treat=="heatmat"&!is.na(LFD)))
model3_MFD_uh<-lmer(MFD~temp_sire+temp_dam+
                      (1|sire)+(1|dam)+(1|sire:dam),
                 subset(data_exp,treat=="unheated"&!is.na(MFD)))
model3_MFD_h<-lmer(MFD~temp_sire+temp_dam+
                     #(1|sire) # Removed to avoid singular fit
                     (1|dam)+(1|sire:dam),
                 subset(data_exp,treat=="heatmat"&!is.na(MFD)))
summary(model3_veg_uh)
summary(model3_veg_h)
summary(model3_FFD_uh)
summary(model3_FFD_h)
summary(model3_LFD_uh)
summary(model3_LFD_h)
summary(model3_MFD_uh)
summary(model3_MFD_h)
```

```{r}
plot(ggpredict(model3_MFD_h,terms=c("temp_sire")))
```

The only significant effect is that of temp_sire on MFD for the heated treatment: the sign of temp_sire goes in the expected direction (later MFD at a given temperature in individuals with sires from warmer soils). 

#### Table 2

```{r}
tab_model(model3_veg_uh,model3_FFD_uh,model3_MFD_uh,model3_LFD_uh,
          transform=NULL,show.intercept=F,show.ci=F,show.se=T,show.stat=T,
          show.r2=T,show.icc=F,show.re.var=F,show.ngroups=F,digits=3,
          dv.labels=c("Vegetative growth","FFD","MFD","LFD"),
          file="output/tables/Table2_uh.html")
tab_model(model3_veg_h,model3_FFD_h,model3_MFD_h,model3_LFD_h,
          transform=NULL,show.intercept=F,show.ci=F,show.se=T,show.stat=T,
          show.r2=T,show.icc=F,show.re.var=F,show.ngroups=F,digits=3,
          dv.labels=c("Vegetative growth","FFD","MFD","LFD"),
          file="output/tables/Table2_h.html")
```

### Proportions of variance

Proportion of variance explained by temperature of origin of sires and dams = proportion of variance explained by fixed effects: marginal R2.

```{r}
PropVar_veg_uh_temp<-as.numeric(r2_nakagawa(model3_veg_uh)$R2_marginal)
PropVar_veg_h_temp<-as.numeric(r2_nakagawa(model3_veg_h)$R2_marginal)
PropVar_FFD_uh_temp<-as.numeric(r2_nakagawa(model3_FFD_uh)$R2_marginal)
PropVar_FFD_h_temp<-as.numeric(r2_nakagawa(model3_FFD_h)$R2_marginal)
PropVar_LFD_uh_temp<-as.numeric(r2_nakagawa(model3_LFD_uh)$R2_marginal)
PropVar_LFD_h_temp<-as.numeric(r2_nakagawa(model3_LFD_h)$R2_marginal)
PropVar_MFD_uh_temp<-as.numeric(r2_nakagawa(model3_MFD_uh)$R2_marginal)
PropVar_MFD_h_temp<-as.numeric(r2_nakagawa(model3_MFD_h)$R2_marginal)
Props_var_temp<-data.frame(
  values=c(PropVar_veg_uh_temp,PropVar_veg_h_temp,
           PropVar_FFD_uh_temp,PropVar_FFD_h_temp,
           PropVar_LFD_uh_temp,PropVar_LFD_h_temp,
           PropVar_MFD_uh_temp,PropVar_MFD_h_temp),
  variable=c("veg","veg","FFD","FFD","LFD","LFD","MFD","MFD"),
  treat=c("unheated","heated","unheated","heated","unheated","heated",
                 "unheated","heated"))
```

```{r}
ggplot(Props_var_temp,aes(x=variable,y=values,fill=treat))+
  geom_bar(stat="identity",position="dodge")
```

##### Temperature dam and sire: partR2

```{r eval=FALSE, include=FALSE}
partR2_model3_veg_uh<-partR2(model3_veg_uh,
                             partvars=c("temp_sire","temp_dam"),
                             R2_type="marginal",nboot=1000)
partR2_model3_veg_h<-partR2(model3_veg_h,
                             partvars=c("temp_sire","temp_dam"),
                             R2_type="marginal",nboot=1000)
partR2_model3_FFD_uh<-partR2(model3_FFD_uh,
                             partvars=c("temp_sire","temp_dam"),
                             R2_type="marginal",nboot=1000)
partR2_model3_FFD_h<-partR2(model3_FFD_h,
                             partvars=c("temp_sire","temp_dam"),
                             R2_type="marginal",nboot=1000)
partR2_model3_LFD_uh<-partR2(model3_LFD_uh,
                             partvars=c("temp_sire","temp_dam"),
                             R2_type="marginal",nboot=1000)
partR2_model3_LFD_h<-partR2(model3_LFD_h,
                             partvars=c("temp_sire","temp_dam"),
                             R2_type="marginal",nboot=1000)
partR2_model3_MFD_uh<-partR2(model3_MFD_uh,
                             partvars=c("temp_sire","temp_dam"),
                             R2_type="marginal",nboot=1000)
partR2_model3_MFD_h<-partR2(model3_MFD_h,
                             partvars=c("temp_sire","temp_dam"),
                             R2_type="marginal",nboot=1000)
save(partR2_model3_veg_uh, file = "output/models/partR2_model3_veg_uh.rda")
save(partR2_model3_veg_h, file = "output/models/partR2_model3_veg_h.rda")
save(partR2_model3_FFD_uh, file = "output/models/partR2_model3_FFD_uh.rda")
save(partR2_model3_FFD_h, file = "output/models/partR2_model3_FFD_h.rda")
save(partR2_model3_LFD_uh, file = "output/models/partR2_model3_LFD_uh.rda")
save(partR2_model3_LFD_h, file = "output/models/partR2_model3_LFD_h.rda")
save(partR2_model3_MFD_uh, file = "output/models/partR2_model3_MFD_uh.rda")
save(partR2_model3_MFD_h, file = "output/models/partR2_model3_MFD_h.rda")
```

```{r}
load("output/models/partR2_model3_veg_uh.rda")
load("output/models/partR2_model3_veg_h.rda")
load("output/models/partR2_model3_FFD_uh.rda")
load("output/models/partR2_model3_FFD_h.rda")
load( "output/models/partR2_model3_LFD_uh.rda")
load("output/models/partR2_model3_LFD_h.rda")
load("output/models/partR2_model3_MFD_uh.rda")
load("output/models/partR2_model3_MFD_h.rda")
```

```{r}
Props_var_temp_mf<-
  rbind(
    partR2_model3_veg_uh$R2%>%mutate(variable="veg",treat="unheated"),
    partR2_model3_veg_h$R2%>%mutate(variable="veg",treat="heated"),
    partR2_model3_FFD_uh$R2%>%mutate(variable="FFD",treat="unheated"),
    partR2_model3_FFD_h$R2%>%mutate(variable="FFD",treat="heated"),
    partR2_model3_LFD_uh$R2%>%mutate(variable="LFD",treat="unheated"),
    partR2_model3_LFD_h$R2%>%mutate(variable="LFD",treat="heated"),
    partR2_model3_MFD_uh$R2%>%mutate(variable="MFD",treat="unheated"),
    partR2_model3_MFD_h$R2%>%mutate(variable="MFD",treat="heated")
    )
```

#### Figure 3

```{r}
fig3<-ggplot()+
  geom_bar(data=Props_var_temp_mf%>%
             filter(term!="temp_sire+temp_dam"&term!="Full")%>%
             mutate(term=factor(term,levels=c("temp_sire","temp_dam"),
                                labels=c("Sire origin temperature",
                                           "Dam origin temperature")),
                    variable=factor(variable,levels=c("veg","FFD","MFD","LFD")),
                    sig=c("n","n","n","n","n","n","n","n",
                          "n","n","n","n","n","n","y","n"))%>%
               mutate(treat=fct_recode(treat,"control"="unheated")),
           aes(x=variable,y=estimate,fill=term),
           stat="identity",color="black",size=0.2)+
  facet_wrap(~factor(treat,levels=c("control","heated")))+
  geom_point(data=Props_var_temp_mf%>%
               filter(term!="temp_sire+temp_dam"&term!="Full")%>%
               mutate(term=factor(term,levels=c("temp_sire","temp_dam"),
                                  labels=c("Sire origin temperature",
                                           "Dam origin temperature")),
                      variable=factor(variable,
                                      levels=c("veg","FFD","MFD","LFD")),
                      sig=c("n","n","n","n","n","n","n","n",
                            "n","n","n","n","n","n","y","n"))%>%
               mutate(treat=fct_recode(treat,"control"="unheated")),
             aes(x=variable,y=estimate,group=term,shape=sig),
             position=position_stack(vjust=.5),show.legend=F,size=2)+
  my_theme_legend()+scale_fill_discrete(type=c("#00BFC4","#F8766D"))+
  scale_shape_manual(values=c(NA,8))+
  xlab(NULL)+ylab("Proportion of variance")+
  theme(legend.position="top",legend.title=element_blank())
fig3
ggsave(file="output/figures/fig3.tiff", plot=fig3,width=5,height=5,dpi=300,
       compression="lzw")
```

### Appendix S5

```{r}
AppS5<-ggplot(data=Props_var_temp_mf%>%
             filter(term!="temp_sire+temp_dam"&term!="Full")%>%
             mutate(term=factor(term,levels=c("temp_sire","temp_dam"),
                                labels=c("Sire origin temperature",
                                           "Dam origin temperature")),
                    variable=factor(variable,levels=c("veg","FFD","MFD","LFD")),
                    sig=c("n","n","n","n","n","n","n","n",
                          "n","n","n","n","n","n","y","n"))%>%
               mutate(treat=fct_recode(treat,"control"="unheated")),
           aes(x=variable,y=estimate,ymin=CI_lower,ymax=CI_upper,color=term))+
  geom_point(size=2,position=position_dodge(0.4),width=0.2)+
  geom_errorbar(position=position_dodge(0.4),width=0.2)+
  facet_wrap(~factor(treat,levels=c("control","heated")))+
  my_theme_legend()+
  scale_color_discrete(type=c("#00BFC4","#F8766D"))+
  xlab(NULL)+ylab("Proportion of variance")+
  theme(legend.position="top",legend.title=element_blank())
AppS5
ggsave(file="output/figures/AppS5.tiff", plot=AppS5,width=7,height=8,dpi=300,
       compression="lzw")
```


### Models mid-parental values

```{r}
data_exp<-data_exp%>%
  mutate(temp_midP=(temp_sire+temp_dam)/2)
```

```{r}
model3_veg_uh_midP<-lmer(mean_sh_growth~temp_midP+
                      (1|sire)+(1|dam),
                      #(1|sire:dam) #R emoved to avoid singular fit
                 subset(data_exp,treat=="unheated"&!is.na(mean_sh_growth)))
model3_veg_h_midP<-lmer(mean_sh_growth~temp_midP+
                     (1|sire)+(1|dam)+(1|sire:dam),
                 subset(data_exp,treat=="heatmat"&!is.na(mean_sh_growth)))
model3_FFD_uh_midP<-lmer(FFD~temp_midP+
                      (1|sire)+(1|dam)+(1|sire:dam),
                 subset(data_exp,treat=="unheated"&!is.na(FFD)))
model3_FFD_h_midP<-lmer(FFD~temp_midP+
                     (1|sire)+(1|dam)+(1|sire:dam),
                 subset(data_exp,treat=="heatmat"&!is.na(FFD)))
model3_LFD_uh_midP<-lmer(LFD~temp_midP+
                      (1|sire)+(1|dam)+(1|sire:dam),
                 subset(data_exp,treat=="unheated"&!is.na(LFD)))
model3_LFD_h_midP<-lmer(LFD~temp_midP+
                     (1|sire)+(1|dam)+(1|sire:dam),
                 subset(data_exp,treat=="heatmat"&!is.na(LFD)))
model3_MFD_uh_midP<-lmer(MFD~temp_midP+
                      (1|sire)+(1|dam)+(1|sire:dam),
                 subset(data_exp,treat=="unheated"&!is.na(MFD)))
model3_MFD_h_midP<-lmer(MFD~temp_midP+
                     (1|sire)+(1|dam)+(1|sire:dam),
                 subset(data_exp,treat=="heatmat"&!is.na(MFD)))
summary(model3_veg_uh_midP)
summary(model3_veg_h_midP)
summary(model3_FFD_uh_midP)
summary(model3_FFD_h_midP)
summary(model3_LFD_uh_midP)
summary(model3_LFD_h_midP)
summary(model3_MFD_uh_midP)
summary(model3_MFD_h_midP)
```

```{r}
plot(ggpredict(model3_MFD_h_midP,terms=c("temp_midP")))
```

#### Appendix S6

```{r}
tab_model(model3_veg_uh_midP,model3_FFD_uh_midP,model3_MFD_uh_midP,
          model3_LFD_uh_midP,
          transform=NULL,show.intercept=F,show.ci=F,show.se=T,show.stat=T,
          show.r2=T,show.icc=F,show.re.var=F,show.ngroups=F,digits=3,
          dv.labels=c("Vegetative growth","FFD","MFD","LFD"),
          file="output/tables/AppS6_uh.html")
tab_model(model3_veg_h_midP,model3_FFD_h_midP,model3_MFD_h_midP,
          model3_LFD_h_midP,
          transform=NULL,show.intercept=F,show.ci=F,show.se=T,show.stat=T,
          show.r2=T,show.icc=F,show.re.var=F,show.ngroups=F,digits=3,
          dv.labels=c("Vegetative growth","FFD","MFD","LFD"),
          file="output/tables/AppS6_h.html")
```

#### Proportions of variance

Proportion of variance explained by temperature of origin of sires and dams = proportion of variance explained by fixed effects: marginal R2.

```{r}
PropVar_veg_uh_temp_midP<-as.numeric(r2_nakagawa(model3_veg_uh_midP,
                                            tolerance=1e-324)$R2_marginal)
PropVar_veg_h_temp_midP<-as.numeric(r2_nakagawa(model3_veg_h_midP)$R2_marginal)
PropVar_FFD_uh_temp_midP<-as.numeric(r2_nakagawa(model3_FFD_uh_midP)$R2_marginal)
PropVar_FFD_h_temp_midP<-as.numeric(r2_nakagawa(model3_FFD_h_midP)$R2_marginal)
PropVar_LFD_uh_temp_midP<-as.numeric(r2_nakagawa(model3_LFD_uh_midP)$R2_marginal)
PropVar_LFD_h_temp_midP<-as.numeric(r2_nakagawa(model3_LFD_h_midP)$R2_marginal)
PropVar_MFD_uh_temp_midP<-as.numeric(r2_nakagawa(model3_MFD_uh_midP)$R2_marginal)
PropVar_MFD_h_temp_midP<-as.numeric(r2_nakagawa(model3_MFD_h_midP)$R2_marginal)
Props_var_temp_midP<-data.frame(
  values=c(PropVar_veg_uh_temp_midP,PropVar_veg_h_temp_midP,
           PropVar_FFD_uh_temp_midP,PropVar_FFD_h_temp_midP,
           PropVar_LFD_uh_temp_midP,PropVar_LFD_h_temp_midP,
           PropVar_MFD_uh_temp_midP,PropVar_MFD_h_temp_midP),
  variable=c("veg","veg","FFD","FFD","LFD","LFD","MFD","MFD"),
  treat=c("unheated","heated","unheated","heated","unheated","heated",
                 "unheated","heated"))
```

```{r}
ggplot(Props_var_temp_midP,aes(x=variable,y=values,fill=treat))+
  geom_bar(stat="identity",position="dodge")
```

## Prediction 4

### Models

```{r}
model4_veg<-lmer(mean_sh_growth~treat*(temp_sire+temp_dam)+
                    (1|sire)+(1|dam)+
                    (1|sire:dam),
                  data_exp)
model4_FFD<-lmer(FFD~treat*(temp_sire+temp_dam)+
                    (1|sire)+(1|dam)+
                    (1|sire:dam),
                  data_exp)
model4_LFD<-lmer(LFD~treat*(temp_sire+temp_dam)+
                    (1|sire)+(1|dam)+
                    (1|sire:dam),
                  data_exp)
model4_MFD<-lmer(MFD~treat*(temp_sire+temp_dam)+
                    (1|sire)+(1|dam)+
                    (1|sire:dam),
                  subset(data_exp,!is.na(MFD)))
summary(model4_veg)
summary(model4_FFD)
summary(model4_LFD)
summary(model4_MFD)
```


Interactions treat:temp_sire and treat:temp_dam never significant. No temperature-related differences among sires and dams in the response to the temperature treatment = No genetic differentiation related to origin temperature in the slope of RNs.

Effect of temp_sire significant in the model for MFD: genetic differentiation in the elevation of RNs related to origin temperature of sire. 

```{r}
plot(ggpredict(model4_MFD,terms=c("temp_sire")))
```

```{r}
grid.arrange(
  grid.arrange(
    plot(ggpredict(model4_veg,terms=c("treat","temp_sire[minmax]")),
         connect_lines=T),
    plot(ggpredict(model4_veg,terms=c("treat","temp_dam[minmax]")),
         connect_lines=T),
    ncol=2),
  grid.arrange(
    plot(ggpredict(model4_FFD,terms=c("treat","temp_sire[minmax]")),
         connect_lines=T),
    plot(ggpredict(model4_FFD,terms=c("treat","temp_dam[minmax]")),
         connect_lines=T),
    ncol=2),
  grid.arrange(
    plot(ggpredict(model4_LFD,terms=c("treat","temp_sire[minmax]")),
         connect_lines=T),
    plot(ggpredict(model4_LFD,terms=c("treat","temp_dam[minmax]")),
         connect_lines=T),
    ncol=2),
  grid.arrange(
    plot(ggpredict(model4_MFD,terms=c("treat","temp_sire[minmax]")),
         connect_lines=T),
    plot(ggpredict(model4_MFD,terms=c("treat","temp_dam[minmax]")),
         connect_lines=T),
    ncol=2),
  ncol=1)
```

#### Table 3

```{r}
tab_model(model4_veg,model4_FFD,model4_MFD,model4_LFD,
          transform=NULL,show.intercept=F,show.ci=F,show.se=T,show.stat=T,
          show.r2=T,show.icc=F,show.re.var=F,show.ngroups=F,digits=3,
          dv.labels=c("Vegetative growth","FFD","MFD","LFD"),
          file="output/tables/Table3.html")
```

#### Figure 4

```{r}
fig4<-ggplot()+
  geom_ribbon(data=data.frame(ggpredict(model3_MFD_h,
                                        terms=c("temp_sire"))),
              aes(x=x,y=predicted,ymin=conf.low,ymax=conf.high),
              fill="red",alpha=0.25)+
  geom_line(data=data.frame(ggpredict(model3_MFD_uh,
                                      terms=c("temp_sire"))),
            aes(x=x,y=predicted),color="blue",linetype="dashed")+
  geom_ribbon(data=data.frame(ggpredict(model3_MFD_uh,
                                        terms=c("temp_sire"))),
              aes(x=x,y=predicted,ymin=conf.low,ymax=conf.high),
              fill="blue",alpha=0.25)+
  geom_line(data=data.frame(ggpredict(model3_MFD_h,
                                      terms=c("temp_sire"))),
            aes(x=x,y=predicted),color="red")+
  geom_ribbon(data=data.frame(ggemmeans(model4_MFD,
                                        terms=c("temp_sire"))),
              aes(x=x,y=predicted,ymin=conf.low,ymax=conf.high),
              fill="darkgrey",alpha=0.25)+
  geom_line(data=data.frame(ggemmeans(model4_MFD,
                                      terms=c("temp_sire"))),
            aes(x=x,y=predicted),color="black",size=0.75)+
  geom_jitter(data=data_exp%>%
                mutate(treat=fct_recode(treat,"control"="unheated")),
              aes(x=temp_sire,y=MFD,color=treat),size=1,alpha=0.1)+
  scale_color_manual(values=c("blue","red"),limits=c("control","heatmat"),
                     labels=c("control","heated"))+
  my_theme_legend()+xlab("Sire origin temperature (ºC)")+
  ylab("Median flowering date")+
  theme(legend.position="top",legend.title=element_blank())+
  theme(legend.position = c(0.25, 0.95),legend.direction="horizontal")+
  theme(legend.background = element_rect(fill = "transparent"),
        legend.spacing.x = unit(0.1, 'cm'))
fig4
ggsave(file="output/figures/fig4.tiff", plot=fig4,width=4,height=4,dpi=300,
       compression="lzw")
```

### Models mid-parental values

```{r}
model4_veg_midP<-lmer(mean_sh_growth~treat*temp_midP+
                    (1|sire)+(1|dam)+
                    (1|sire:dam),
                  data_exp)
model4_FFD_midP<-lmer(FFD~treat*temp_midP+
                    (1|sire)+(1|dam)+
                    (1|sire:dam),
                  data_exp)
model4_LFD_midP<-lmer(LFD~treat*temp_midP+
                    (1|sire)+(1|dam)+
                    (1|sire:dam),
                  data_exp)
model4_MFD_midP<-lmer(MFD~treat*temp_midP+
                    (1|sire)+(1|dam)+
                    (1|sire:dam),
                  data_exp)
summary(model4_veg_midP)
summary(model4_FFD_midP)
summary(model4_LFD_midP)
summary(model4_MFD_midP)
```

```{r}
plot(ggpredict(model4_MFD_midP,terms=c("temp_midP")))
```

#### Appendix S7

```{r}
tab_model(model4_veg_midP,model4_FFD_midP,model4_MFD_midP,model4_LFD_midP,
          transform=NULL,show.intercept=F,show.ci=F,show.se=T,show.stat=T,
          show.r2=T,show.icc=F,show.re.var=F,show.ngroups=F,digits=3,
          dv.labels=c("Vegetative growth","FFD","MFD","LFD"),
          file="output/tables/AppS7.html")
```

# Prediction 3 with breeding values (elevation)

Get breeding values/BLUPs (elevation of RNs) from models in Prediction 1.

What are these breeding values? Their mean is approx. zero. They represent the difference between that genotype's predicted response and the (population-level) average predicted response. They indicate the difference in gentoype phenology relative to the population-average, so more positive values of BLUP intercept indicate that the genotype's response occurs above the population-level average, and  negative values are below the population-level average. For vegetative growth, positive values indicate that growth is higher than the population-level average, and negative values indicate that growth is lower than the population-level average. For FFD/MFD/LFD, positive values indicate that phenology is later than the population-level average, and negative values indicate that phenology is earlier than the population-level average.  

The population-level averages are represented by the intercepts of the models

```{r}
averages_elev<-rbind(
  data.frame(average=as.numeric(fixef(model1a_veg_uh)))%>%
    mutate(name="veg_uh_elev"),
  data.frame(average=as.numeric(fixef(model1a_veg_h)))%>%
    mutate(name="veg_h_elev"),
  data.frame(average=as.numeric(fixef(model1a_FFD_uh)))%>%
    mutate(name="FFD_uh_elev"),
  data.frame(average=as.numeric(fixef(model1a_FFD_h)))%>%
    mutate(name="FFD_h_elev"),
  data.frame(average=as.numeric(fixef(model1a_MFD_uh)))%>%
    mutate(name="MFD_uh_elev"),
  data.frame(average=as.numeric(fixef(model1a_MFD_h)))%>%
    mutate(name="MFD_h_elev"),
  data.frame(average=as.numeric(fixef(model1a_LFD_uh)))%>%
    mutate(name="LFD_uh_elev"),
  data.frame(average=as.numeric(fixef(model1a_LFD_h)))%>%
    mutate(name="LFD_h_elev")
  )%>%
  separate(name,into=c("var","treat","type"))
```

```{r}
breedV_elev_sires<-left_join(
  left_join(
    left_join(
      left_join(
        left_join(
          left_join(
            left_join(
              ranef(model1a_veg_uh)$sire%>%rownames_to_column(var="sire")%>%
                rename(breedV_veg_uh_elev_sire="(Intercept)"),
              ranef(model1a_veg_h)$sire%>%rownames_to_column(var="sire")%>%
                rename(breedV_veg_h_elev_sire="(Intercept)")),
            ranef(model1a_FFD_uh)$sire%>%rownames_to_column(var="sire")%>%
              rename(breedV_FFD_uh_elev_sire="(Intercept)")
            ),
          ranef(model1a_FFD_h)$sire%>%rownames_to_column(var="sire")%>%
            rename(breedV_FFD_h_elev_sire="(Intercept)")
          ),
        ranef(model1a_MFD_uh)$sire%>%rownames_to_column(var="sire")%>%
          rename(breedV_MFD_uh_elev_sire="(Intercept)")
        ),
      ranef(model1a_MFD_h)$sire%>%rownames_to_column(var="sire")%>%
        rename(breedV_MFD_h_elev_sire="(Intercept)")
      ),
    ranef(model1a_LFD_uh)$sire%>%rownames_to_column(var="sire")%>%
      rename(breedV_LFD_uh_elev_sire="(Intercept)")
    ),
  ranef(model1a_LFD_h)$sire%>%rownames_to_column(var="sire")%>%
    rename(breedV_LFD_h_elev_sire="(Intercept)")
  )%>%
  pivot_longer(cols=breedV_veg_uh_elev_sire:breedV_LFD_h_elev_sire,
               names_to=c("remove","var","treat","type","comp"),names_sep="_",
               values_to="breedV")%>%
  dplyr::select(-remove)%>%
  left_join(averages_elev)%>%mutate(breedV_av=breedV+average)
# Add BLUP intercept estimates to the population-level average
```

```{r}
breedV_elev_dams<-left_join(
  left_join(
    left_join(
      left_join(
        left_join(
          left_join(
            left_join(
              ranef(model1a_veg_uh)$dam%>%rownames_to_column(var="dam")%>%
                rename(breedV_veg_uh_elev_dam="(Intercept)"),
              ranef(model1a_veg_h)$dam%>%rownames_to_column(var="dam")%>%
                rename(breedV_veg_h_elev_dam="(Intercept)")),
            ranef(model1a_FFD_uh)$dam%>%rownames_to_column(var="dam")%>%
              rename(breedV_FFD_uh_elev_dam="(Intercept)")
            ),
          ranef(model1a_FFD_h)$dam%>%rownames_to_column(var="dam")%>%
            rename(breedV_FFD_h_elev_dam="(Intercept)")
          ),
        ranef(model1a_MFD_uh)$dam%>%rownames_to_column(var="dam")%>%
          rename(breedV_MFD_uh_elev_dam="(Intercept)")
        ),
      ranef(model1a_MFD_h)$dam%>%rownames_to_column(var="dam")%>%
        rename(breedV_MFD_h_elev_dam="(Intercept)")
      ),
    ranef(model1a_LFD_uh)$dam%>%rownames_to_column(var="dam")%>%
      rename(breedV_LFD_uh_elev_dam="(Intercept)")
    ),
  ranef(model1a_LFD_h)$dam%>%rownames_to_column(var="dam")%>%
    rename(breedV_LFD_h_elev_dam="(Intercept)")
  )%>%
  pivot_longer(cols=breedV_veg_uh_elev_dam:breedV_LFD_h_elev_dam,
               names_to=c("remove","var","treat","type","comp"),names_sep="_",
               values_to="breedV")%>%
  dplyr::select(-remove)%>%
  left_join(averages_elev)%>%mutate(breedV_av=breedV+average)
# Add BLUP intercept estimates to the population-level average
```

Plot BLUPs for sires for the elevation

```{r}
ggplot(breedV_elev_sires, aes(sire, breedV))+
  geom_point(aes(group = sire, colour = sire), size = 4) +
  ylab("BLUP elevation")+
  geom_hline(yintercept = 0, lty = 2) + 
  facet_grid(rows=vars(var),cols=vars(treat))+
  theme(legend.position="none")
```

Plot BLUPs for dams for the elevation

```{r}
ggplot(breedV_elev_dams, aes(dam, breedV))+
geom_point(aes(group = dam, colour = dam), size = 4) +
  ylab("BLUP elevation")+
  geom_hline(yintercept = 0, lty = 2) + 
  facet_grid(rows=vars(var),cols=vars(treat))+
  theme(legend.position="none")
```

Plot BLUPs for sires + population level averages for the elevation

```{r}
ggplot()+
  geom_line(data=breedV_elev_sires%>%filter(var=="veg"),
            aes(x=treat,y=breedV_av,color=sire,group=sire),size=1,alpha=0.5)+
  geom_point(data=breedV_elev_sires%>%filter(var=="veg"),
             aes(x=treat,y=breedV_av,color=sire,group=sire),size=5,alpha=0.5)+
  geom_point(data=breedV_elev_sires%>%filter(var=="veg"),
             aes(x=treat,y=average,group=sire),
             color="black",size=5,shape=15)+
  geom_line(data=breedV_elev_sires%>%filter(var=="veg"),
            aes(x=treat,y=average,group=sire),
            color="black",size=1,linetype="dashed")+
  ylab("BLUP elevation + population-level average")+ggtitle("veg")
ggplot()+
  geom_line(data=breedV_elev_sires%>%filter(var=="FFD"),
            aes(x=treat,y=breedV_av,color=sire,group=sire),size=1,alpha=0.5)+
  geom_point(data=breedV_elev_sires%>%filter(var=="FFD"),
             aes(x=treat,y=breedV_av,color=sire,group=sire),size=5,alpha=0.5)+
  geom_point(data=breedV_elev_sires%>%filter(var=="FFD"),
             aes(x=treat,y=average,group=sire),
             color="black",size=5,shape=15)+
  geom_line(data=breedV_elev_sires%>%filter(var=="FFD"),
            aes(x=treat,y=average,group=sire),
            color="black",size=1,linetype="dashed")+
  ylab("BLUP elevation + population-level average")+ggtitle("FFD")
ggplot()+
  geom_line(data=breedV_elev_sires%>%filter(var=="MFD"),
            aes(x=treat,y=breedV_av,color=sire,group=sire),size=1,alpha=0.5)+
  geom_point(data=breedV_elev_sires%>%filter(var=="MFD"),
             aes(x=treat,y=breedV_av,color=sire,group=sire),size=5,alpha=0.5)+
  geom_point(data=breedV_elev_sires%>%filter(var=="MFD"),
             aes(x=treat,y=average,group=sire),
             color="black",size=5,shape=15)+
  geom_line(data=breedV_elev_sires%>%filter(var=="MFD"),
            aes(x=treat,y=average,group=sire),
            color="black",size=1,linetype="dashed")+
  ylab("BLUP elevation + population-level average")+ggtitle("MFD")
ggplot()+
  geom_line(data=breedV_elev_sires%>%filter(var=="LFD"),
            aes(x=treat,y=breedV_av,color=sire,group=sire),size=1,alpha=0.5)+
  geom_point(data=breedV_elev_sires%>%filter(var=="LFD"),
             aes(x=treat,y=breedV_av,color=sire,group=sire),size=5,alpha=0.5)+
  geom_point(data=breedV_elev_sires%>%filter(var=="LFD"),
             aes(x=treat,y=average,group=sire),
             color="black",size=5,shape=15)+
  geom_line(data=breedV_elev_sires%>%filter(var=="LFD"),
            aes(x=treat,y=average,group=sire),
            color="black",size=1,linetype="dashed")+
  ylab("BLUP elevation + population-level average")+ggtitle("LFD")
```

Plot BLUPs for dams + population level averages for the elevation

```{r}
ggplot()+
  geom_line(data=breedV_elev_dams%>%filter(var=="veg"),
            aes(x=treat,y=breedV_av,color=dam,group=dam),size=1,alpha=0.5)+
  geom_point(data=breedV_elev_dams%>%filter(var=="veg"),
             aes(x=treat,y=breedV_av,color=dam,group=dam),size=5,alpha=0.5)+
  geom_point(data=breedV_elev_dams%>%filter(var=="veg"),
             aes(x=treat,y=average,group=dam),
             color="black",size=5,shape=15)+
  geom_line(data=breedV_elev_dams%>%filter(var=="veg"),
            aes(x=treat,y=average,group=dam),
            color="black",size=1,linetype="dashed")+
  ylab("BLUP elevation - population-level average")+ggtitle("veg")
ggplot()+
  geom_line(data=breedV_elev_dams%>%filter(var=="FFD"),
            aes(x=treat,y=breedV_av,color=dam,group=dam),size=1,alpha=0.5)+
  geom_point(data=breedV_elev_dams%>%filter(var=="FFD"),
             aes(x=treat,y=breedV_av,color=dam,group=dam),size=5,alpha=0.5)+
  geom_point(data=breedV_elev_dams%>%filter(var=="FFD"),
             aes(x=treat,y=average,group=dam),
             color="black",size=5,shape=15)+
  geom_line(data=breedV_elev_dams%>%filter(var=="FFD"),
            aes(x=treat,y=average,group=dam),
            color="black",size=1,linetype="dashed")+
  ylab("BLUP elevation - population-level average")+ggtitle("FFD")
ggplot()+
  geom_line(data=breedV_elev_dams%>%filter(var=="MFD"),
            aes(x=treat,y=breedV_av,color=dam,group=dam),size=1,alpha=0.5)+
  geom_point(data=breedV_elev_dams%>%filter(var=="MFD"),
             aes(x=treat,y=breedV_av,color=dam,group=dam),size=5,alpha=0.5)+
  geom_point(data=breedV_elev_dams%>%filter(var=="MFD"),
             aes(x=treat,y=average,group=dam),
             color="black",size=5,shape=15)+
  geom_line(data=breedV_elev_dams%>%filter(var=="MFD"),
            aes(x=treat,y=average,group=dam),
            color="black",size=1,linetype="dashed")+
  ylab("BLUP elevation - population-level average")+ggtitle("MFD")
ggplot()+
  geom_line(data=breedV_elev_dams%>%filter(var=="LFD"),
            aes(x=treat,y=breedV_av,color=dam,group=dam),size=1,alpha=0.5)+
  geom_point(data=breedV_elev_dams%>%filter(var=="LFD"),
             aes(x=treat,y=breedV_av,color=dam,group=dam),size=5,alpha=0.5)+
  geom_point(data=breedV_elev_dams%>%filter(var=="LFD"),
             aes(x=treat,y=average,group=dam),
             color="black",size=5,shape=15)+
  geom_line(data=breedV_elev_dams%>%filter(var=="LFD"),
            aes(x=treat,y=average,group=dam),
            color="black",size=1,linetype="dashed")+
  ylab("BLUP elevation - population-level average")+ggtitle("LFD")
```

Add origin temperatures:

```{r}
breedV_elev_sires<-breedV_elev_sires%>%
  left_join(data_exp%>%dplyr::select(sire,temp_sire)%>%unique())
breedV_elev_dams<-breedV_elev_dams%>%
  left_join(data_exp%>%dplyr::select(dam,temp_dam)%>%unique())
```

## Sire breeding values ~ origin temperatures

```{r}
mod_bV_elev_temp_sire_veg_uh<-lm(breedV~temp_sire,
                                 subset(breedV_elev_sires,
                                        var=="veg"&treat=="uh"))
mod_bV_elev_temp_sire_veg_h<-lm(breedV~temp_sire,
                                 subset(breedV_elev_sires,
                                        var=="veg"&treat=="h"))
mod_bV_elev_temp_sire_FFD_uh<-lm(breedV~temp_sire,
                                 subset(breedV_elev_sires,
                                        var=="FFD"&treat=="uh"))
mod_bV_elev_temp_sire_FFD_h<-lm(breedV~temp_sire,
                                 subset(breedV_elev_sires,
                                        var=="FFD"&treat=="h"))
mod_bV_elev_temp_sire_MFD_uh<-lm(breedV~temp_sire,
                                 subset(breedV_elev_sires,
                                        var=="MFD"&treat=="uh"))
mod_bV_elev_temp_sire_MFD_h<-lm(breedV~temp_sire,
                                 subset(breedV_elev_sires,
                                        var=="MFD"&treat=="h"))
mod_bV_elev_temp_sire_LFD_uh<-lm(breedV~temp_sire,
                                 subset(breedV_elev_sires,
                                        var=="LFD"&treat=="uh"))
mod_bV_elev_temp_sire_LFD_h<-lm(breedV~temp_sire,
                                 subset(breedV_elev_sires,
                                        var=="LFD"&treat=="h"))
```

```{r}
summary(mod_bV_elev_temp_sire_veg_uh)
summary(mod_bV_elev_temp_sire_veg_h)
summary(mod_bV_elev_temp_sire_FFD_uh)
summary(mod_bV_elev_temp_sire_FFD_h)
summary(mod_bV_elev_temp_sire_MFD_uh)
summary(mod_bV_elev_temp_sire_MFD_h) #*
summary(mod_bV_elev_temp_sire_LFD_uh)
summary(mod_bV_elev_temp_sire_LFD_h)
```

### Plots

```{r}
ggplot(breedV_elev_sires,aes(x=temp_sire,y=breedV_av))+
  geom_point()+geom_smooth(method="lm")+
  facet_grid2(var~treat,scales="free",independent="all")+
  xlab("Sire origin temperature")+
  ylab("Sire breeding value for elevation + population-level average")
```

#### Figure S4A

Only significant relationship between sire breeding value for elevation and origin temperature occurs for MFD in the heated treatment (agrees with what we found before!)

```{r}
figS4A<-plot(ggpredict(mod_bV_elev_temp_sire_MFD_h,terms="temp_sire"),add.data=T)+
  xlab("Sire origin temperature (ºC)")+
  ylab("Sire breeding value for the\nresponse of median flowering date")+
  my_theme()+ggtitle(NULL)
figS4A
ggsave(file="output/figures/figS4A.tiff", plot=figS4A,width=5,height=4,dpi=300,
       compression="lzw")
```

## Dam breeding values ~ origin temperatures

```{r}
mod_bV_elev_temp_dam_veg_uh<-lm(breedV~temp_dam,
                                 subset(breedV_elev_dams,
                                        var=="veg"&treat=="uh"))
mod_bV_elev_temp_dam_veg_h<-lm(breedV~temp_dam,
                                 subset(breedV_elev_dams,
                                        var=="veg"&treat=="h"))
mod_bV_elev_temp_dam_FFD_uh<-lm(breedV~temp_dam,
                                 subset(breedV_elev_dams,
                                        var=="FFD"&treat=="uh"))
mod_bV_elev_temp_dam_FFD_h<-lm(breedV~temp_dam,
                                 subset(breedV_elev_dams,
                                        var=="FFD"&treat=="h"))
mod_bV_elev_temp_dam_MFD_uh<-lm(breedV~temp_dam,
                                 subset(breedV_elev_dams,
                                        var=="MFD"&treat=="uh"))
mod_bV_elev_temp_dam_MFD_h<-lm(breedV~temp_dam,
                                 subset(breedV_elev_dams,
                                        var=="MFD"&treat=="h"))
mod_bV_elev_temp_dam_LFD_uh<-lm(breedV~temp_dam,
                                 subset(breedV_elev_dams,
                                        var=="LFD"&treat=="uh"))
mod_bV_elev_temp_dam_LFD_h<-lm(breedV~temp_dam,
                                 subset(breedV_elev_dams,
                                        var=="LFD"&treat=="h"))
```

```{r}
summary(mod_bV_elev_temp_dam_veg_uh)
summary(mod_bV_elev_temp_dam_veg_h)
summary(mod_bV_elev_temp_dam_FFD_uh)
summary(mod_bV_elev_temp_dam_FFD_h)
summary(mod_bV_elev_temp_dam_MFD_uh)
summary(mod_bV_elev_temp_dam_MFD_h)
summary(mod_bV_elev_temp_dam_LFD_uh)
summary(mod_bV_elev_temp_dam_LFD_h)
# All NS
```

### Plots

```{r}
ggplot(breedV_elev_dams,aes(x=temp_dam,y=breedV_av))+
  geom_point()+geom_smooth(method="lm")+
  facet_grid2(var~treat,scales="free",independent="all")+
  xlab("Dam origin temperature")+
  ylab("Dam breeding value for elevation + population-level average")
```

# Prediction 4 with breeding values (slope)

Get breeding values/BLUPs (slope of RNs) from models in Prediction 2.

What are these breeding values? Their mean is approx. zero. They represent the difference in slope (relative steepness of change) between the population-level average response and the response of the individual genotype. They indicate the difference in steepness of change in phenology relative to the population-average steepness. In this case, the difference in phenology between the heated and unheated treatment relative to the population-average difference. So more positive values of BLUP slope indicate that the genotype's difference is larger than the population-level average difference, and more negative values of BLUP slope indicate that the genotype's difference is smaller than the population-level average difference. 

The population-level average differences are represented by the fixed effects of treatment in the models

```{r}
averages_slope<-rbind(
  data.frame(ggpredict(model1b_veg))[1:2]%>%
    pivot_wider(names_from=treat.x,values_from=treat.predicted)%>%
    mutate(slope=heatmat-unheated,name="veg_slope"), # slope = heated - unheated
  data.frame(ggpredict(model1b_FFD))[1:2]%>%
    pivot_wider(names_from=treat.x,values_from=treat.predicted)%>%
    mutate(slope=heatmat-unheated,name="FFD_slope"),
  data.frame(ggpredict(model1b_MFD))[1:2]%>%
    pivot_wider(names_from=treat.x,values_from=treat.predicted)%>%
    mutate(slope=heatmat-unheated,name="MFD_slope"),
  data.frame(ggpredict(model1b_LFD))[1:2]%>%
    pivot_wider(names_from=treat.x,values_from=treat.predicted)%>%
    mutate(slope=heatmat-unheated,name="LFD_slope")
  )%>%
  separate(name,into=c("var","type"))%>%
  rename(average=slope,av_heatmat=heatmat,av_unheated=unheated)%>%
  dplyr::select(var,type,average,av_heatmat,av_unheated)
```

```{r}
breedV_slope_sires<-left_join(
  ranef(model1b_FFD)$`treat:sire`%>%
    rownames_to_column(var="treat_sire")%>%
    rename(breedV_FFD_slope_sire="(Intercept)")%>%
    separate(treat_sire,into=c("treat","sire"))%>%
    pivot_wider(names_from="treat",
                values_from="breedV_FFD_slope_sire",
                names_prefix="breedV_FFD_slope_sire_"),
  ranef(model1b_MFD)$`treat:sire`%>%
    rownames_to_column(var="treat_sire")%>%
    rename(breedV_MFD_slope_sire="(Intercept)")%>%
    separate(treat_sire,into=c("treat","sire"))%>%
    pivot_wider(names_from="treat",
                values_from="breedV_MFD_slope_sire",
                names_prefix="breedV_MFD_slope_sire_")
  )%>%
  pivot_longer(cols=breedV_FFD_slope_sire_heatmat:breedV_MFD_slope_sire_unheated,
               names_to=c("remove","var","type","comp","treat"),names_sep="_",
               values_to="breedV")%>%
  dplyr::select(-remove)%>%
  pivot_wider(names_from="treat",values_from="breedV",names_prefix="breedV_")%>%
  mutate(breedV=breedV_heatmat-breedV_unheated)%>% # Difference heated-unheated
  left_join(averages_slope)%>%mutate(breedV_av=average+breedV)
# Add BLUP slope estimates to the population-level average
```

```{r}
breedV_slope_dams<-left_join(
  ranef(model1b_veg)$`treat:dam`%>%
    rownames_to_column(var="treat_dam")%>%
    rename(breedV_veg_slope_dam="(Intercept)")%>%
    separate(treat_dam,into=c("treat","dam"))%>%
    pivot_wider(names_from="treat",
                values_from="breedV_veg_slope_dam",
                names_prefix="breedV_veg_slope_dam_"),
  ranef(model1b_LFD)$`treat:dam`%>%
    rownames_to_column(var="treat_dam")%>%
    rename(breedV_LFD_slope_dam="(Intercept)")%>%
    separate(treat_dam,into=c("treat","dam"))%>%
    pivot_wider(names_from="treat",
                values_from="breedV_LFD_slope_dam",
                names_prefix="breedV_LFD_slope_dam_")
  )%>%
    pivot_longer(cols=breedV_veg_slope_dam_heatmat:breedV_LFD_slope_dam_unheated,
               names_to=c("remove","var","type","comp","treat"),names_sep="_",
               values_to="breedV")%>%
  dplyr::select(-remove)%>%
  pivot_wider(names_from="treat",values_from="breedV",names_prefix="breedV_")%>%
  mutate(breedV=breedV_heatmat-breedV_unheated)%>% # Difference heated-unheated
  left_join(averages_slope)%>%mutate(breedV_av=average+breedV)
# Add BLUP slope estimates to the population-level average
```

Plot BLUPs for sires for the elevation

```{r}
ggplot(breedV_slope_sires, aes(sire, breedV))+
geom_point(aes(group = sire, colour = sire), size = 4) +
  ylab("BLUP slope")+
  geom_hline(yintercept = 0, lty = 2) + 
  facet_wrap(~var)+
  theme(legend.position="none")
```

Plot BLUPs for dams for the elevation

```{r}
ggplot(breedV_slope_dams, aes(dam, breedV))+
geom_point(aes(group = dam, colour = dam), size = 4) +
  ylab("BLUP slope")+
  geom_hline(yintercept = 0, lty = 2) + 
  facet_wrap(~var)+
  theme(legend.position="none")
```

Plot BLUPs for sires + population level averages for the slope

```{r}
breedV_slope_sires%>%
  dplyr::select(sire,var,breedV_heatmat,breedV_unheated,av_heatmat,av_unheated)%>%
  pivot_longer(cols=breedV_heatmat:av_unheated,names_to="treat",values_to="values")%>%
  separate(treat,into=c("type","treat"))%>%
  pivot_wider(names_from="type",values_from="values")%>%
  filter(var=="FFD")%>%
  ggplot()+
  geom_line(aes(x=treat,y=breedV+av,color=sire,group=sire),size=1,alpha=0.5)+
  geom_point(aes(x=treat,y=breedV+av,color=sire,group=sire),size=5,alpha=0.5)+
  geom_point(aes(x=treat,y=av,group=sire),color="black",size=5,shape=15)+
  geom_line(aes(x=treat,y=av,group=sire),color="black",size=1,linetype="dashed")+
  ylab("BLUP slope + population-level average")+ggtitle("FFD")
breedV_slope_sires%>%
  dplyr::select(sire,var,breedV_heatmat,breedV_unheated,av_heatmat,av_unheated)%>%
  pivot_longer(cols=breedV_heatmat:av_unheated,names_to="treat",values_to="values")%>%
  separate(treat,into=c("type","treat"))%>%
  pivot_wider(names_from="type",values_from="values")%>%
  filter(var=="MFD")%>%
  ggplot()+
  geom_line(aes(x=treat,y=breedV+av,color=sire,group=sire),size=1,alpha=0.5)+
  geom_point(aes(x=treat,y=breedV+av,color=sire,group=sire),size=5,alpha=0.5)+
  geom_point(aes(x=treat,y=av,group=sire),color="black",size=5,shape=15)+
  geom_line(aes(x=treat,y=av,group=sire),color="black",size=1,linetype="dashed")+
  ylab("BLUP slope + population-level average")+ggtitle("MFD")
```

Plot BLUPs for dams + population level averages for the slope

```{r}
breedV_slope_dams%>%
  dplyr::select(dam,var,breedV_heatmat,breedV_unheated,av_heatmat,av_unheated)%>%
  pivot_longer(cols=breedV_heatmat:av_unheated,names_to="treat",values_to="values")%>%
  separate(treat,into=c("type","treat"))%>%
  pivot_wider(names_from="type",values_from="values")%>%
  filter(var=="veg")%>%
  ggplot()+
  geom_line(aes(x=treat,y=breedV+av,color=dam,group=dam),size=1,alpha=0.5)+
  geom_point(aes(x=treat,y=breedV+av,color=dam,group=dam),size=5,alpha=0.5)+
  geom_point(aes(x=treat,y=av,group=dam),color="black",size=5,shape=15)+
  geom_line(aes(x=treat,y=av,group=dam),color="black",size=1,linetype="dashed")+
  ylab("BLUP slope + population-level average")+ggtitle("veg")
breedV_slope_dams%>%
  dplyr::select(dam,var,breedV_heatmat,breedV_unheated,av_heatmat,av_unheated)%>%
  pivot_longer(cols=breedV_heatmat:av_unheated,names_to="treat",values_to="values")%>%
  separate(treat,into=c("type","treat"))%>%
  pivot_wider(names_from="type",values_from="values")%>%
  filter(var=="LFD")%>%
  ggplot()+
  geom_line(aes(x=treat,y=breedV+av,color=dam,group=dam),size=1,alpha=0.5)+
  geom_point(aes(x=treat,y=breedV+av,color=dam,group=dam),size=5,alpha=0.5)+
  geom_point(aes(x=treat,y=av,group=dam),color="black",size=5,shape=15)+
  geom_line(aes(x=treat,y=av,group=dam),color="black",size=1,linetype="dashed")+
  ylab("BLUP slope + population-level average")+ggtitle("LFD")
```

Add origin temperatures:

```{r}
breedV_slope_sires<-breedV_slope_sires%>%
  left_join(data_exp%>%dplyr::select(sire,temp_sire)%>%unique())
breedV_slope_dams<-breedV_slope_dams%>%
  left_join(data_exp%>%dplyr::select(dam,temp_dam)%>%unique())
```

## Sire breeding values ~ origin temperatures

```{r}
mod_bV_slope_temp_sire_FFD<-lm(breedV~temp_sire,
                                 subset(breedV_slope_sires,var=="FFD"))
mod_bV_slope_temp_sire_MFD<-lm(breedV~temp_sire,
                                 subset(breedV_slope_sires,var=="MFD"))
```

```{r}
summary(mod_bV_slope_temp_sire_FFD)
summary(mod_bV_slope_temp_sire_MFD)
# All NS
```

### Plots

```{r}
ggplot(breedV_slope_sires,aes(x=temp_sire,y=breedV_av))+
  geom_point()+geom_smooth(method="lm")+
  facet_wrap(~var,scales="free")+
  xlab("Sire origin temperature")+
  ylab("Sire breeding value for slope + population-level average")
```

## Dam breeding values ~ origin temperatures

```{r}
mod_bV_slope_temp_dam_veg<-lm(breedV~temp_dam,
                               subset(breedV_slope_dams,var=="veg"))
mod_bV_slope_temp_dam_LFD<-lm(breedV~temp_dam,
                                 subset(breedV_slope_dams,var=="LFD"))
```

```{r}
summary(mod_bV_slope_temp_dam_veg)
summary(mod_bV_slope_temp_dam_LFD)
# All NS
```

### Plots

```{r}
ggplot(breedV_slope_dams,aes(x=temp_dam,y=breedV_av))+
  geom_point()+geom_smooth(method="lm")+
  facet_wrap(~var,scales="free")+
  xlab("Dam origin temperature")+
  ylab("Dam breeding value for slope + population-level average")
```

# Rate of evolution

R=h2*Beta

Using a modification of Breeder's equation, combine heritability values from this study with selection gradients from the Ecology paper to calculate a measure of the response to selection. We use selection gradients and not selection differentials, thus assuming that selection acts only on phenology.

We can only do this for FFD, as selection gradients are only available for FFD.

Read selection gradients for different temperatures from Ecology paper, and sd for FFD in 2018:

```{r}
selgrads_temp<-read_csv("data/clean/selgrads_temp_from_Ecology_paper.csv")
sd_FFD_2018_from_Ecology_paper<-read_csv("data/clean/sd_FFD_2018_from_Ecology_paper.csv")
```

Calculate R (Response to selection):

```{r}
R_UH<-her_FFD_uh* # Heritability for UH treatment
  (selgrads_temp%>%filter(group==14))$slope* # Selection gradient at 14ºC
  sd_FFD_2018_from_Ecology_paper$sd_FFD_2018 # sd for FFD in 2018
R_H<-her_FFD_h* # Heritability for H treatment
  (selgrads_temp%>%filter(group==21))$slope* # Selection gradient at 21ºC
  sd_FFD_2018_from_Ecology_paper$sd_FFD_2018 # sd for FFD in 2018
R_UH
R_H
```

Calculate R at 8ºC and at 27ºC (more similar to unheated/heated conditions in the field):

```{r}
her_FFD_uh* # Heritability for UH treatment
  (selgrads_temp%>%filter(group==8))$slope* # Selection gradient at 8ºC
  sd_FFD_2018_from_Ecology_paper$sd_FFD_2018 # sd for FFD in 2018
her_FFD_h* # Heritability for H treatment
  (selgrads_temp%>%filter(group==27))$slope* # Selection gradient at 27ºC
  sd_FFD_2018_from_Ecology_paper$sd_FFD_2018 # sd for FFD in 2018
```

# Session info

```{r}
sessionInfo()
```
