==> 5_analyses_20241127.Rmd <==
---
title: Small-scale genetic differentiation, in mean flowering time, but not in plasticity, along a geothermal heating gradient
subtitle: Data analyses
author: "Alicia Valdés"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output:
  html_document:
    toc: yes
    toc_depth: '4'
    df_print: paged
  pdf_document:
    toc: yes
    toc_depth: 4
  word_document:
    toc: yes
    toc_depth: '4'
editor_options:
  chunk_output_type: inline
---

```{r setup, include=FALSE}
options(tibble.width = Inf)
knitr::opts_chunk$set(warning = FALSE, message = FALSE) 
```

# Load previously created objects

```{r}
load("output/models/partR2_model3_veg_uh.rda")
load("output/models/partR2_model3_veg_h.rda")
load("output/models/partR2_model3_FFD_uh.rda")
load("output/models/partR2_model3_FFD_h.rda")
load( "output/models/partR2_model3_LFD_uh.rda")
load("output/models/partR2_model3_LFD_h.rda")
load("output/models/partR2_model3_MFD_uh.rda")
load("output/models/partR2_model3_MFD_h.rda")
```

# Load the packages used

```{r load packages, include=FALSE}
library(tidyverse)
library(readxl)
library(ggeffects)
library(knitr)
library(lme4)
library(lmerTest)
library(gridExtra)
library(ggthemes)
library(sjPlot)
library(cowplot)
library(performance)
library(partR2)
library(ggh4x)
```

# Create custom ggplot themes

```{r Define ggplot themes and palettes, include=FALSE}
my_theme <- function(){
  theme_base()+theme(plot.background=element_rect(fill="white", colour=NA))+
  theme(legend.position="none")+theme(text=element_text(family="serif"))+
  theme(plot.title = element_text(hjust =-0.06))
}
my_theme_legend <- function(){
  theme_base()+theme(plot.background=element_rect(fill="white", colour=NA))+
  theme(text=element_text(family="serif"))+
  theme(plot.title = element_text(hjust =-0.06))
}
```

# Read data from csv files

```{r}
data_exp<-read_csv(file="data/clean/data_exp.csv")%>%
  mutate(across(c(unique_id, treat,sire,dam,dam_sire),as.factor),
         across(c(total_n_fl,FFD,LFD,MFD),as.integer),
         across(c(temp_dam,temp_sire,mean_sh_growth,mean_sh_h_date1,
                  mean_sh_h_date2),as.double))
```

# Change incorrect temperature

Temperature for dam 28 should be 45.5, not 25.5.

```{r}
data_exp <- data_exp %>%
  mutate(temp_dam = if_else(dam == 28, 45.5, temp_dam))
```

# Check that temperatures for sires and dams coincide with temperatures in list of sires and dams:

```{r}
temps_dams_sires<-data_exp%>%
  dplyr::select(dam,sire,temp_dam,temp_sire)%>%
  distinct()%>% # Removes duplicate rows
  pivot_longer(cols=starts_with("temp"),names_to="type",values_to="temp")%>%
  mutate(id = case_when(type=="temp_dam"~paste0("D",dam),
                        type=="temp_sire"~paste0("S",sire)))%>%
  dplyr::select(id,temp)
# Read csv file with temperatures from sire and dam list
temps_dams_sires_list<-read_csv("data/clean/temps_dams_sires_list.csv")
# Merge the data frames on the "id" column
merged_data<-temps_dams_sires%>%
  inner_join(temps_dams_sires_list,by="id",suffix=c("_data","_list"))
# Check for discrepancies
discrepancies<-merged_data%>%filter(temp_data!=temp_list)
# View the discrepancies
print(discrepancies)  # No discrepancies
```

# Remove 2 cases that are "out" of the NCII blocks

```{r}
data_exp<-data_exp%>%filter(!(dam==26&sire==260))
```

# Data analyses

## Additive genetic variation in flowering time

### Models

```{r}
model1_FFD_uh<-lmer(FFD~1+(1|sire)+(1|dam)+(1|sire:dam),
                 subset(data_exp,treat=="unheated"))
model1_FFD_h<-lmer(FFD~1+(1|sire)+(1|dam)+(1|sire:dam),
                 subset(data_exp,treat=="heatmat"))
model1_LFD_uh<-lmer(LFD~1+(1|sire)+(1|dam)+(1|sire:dam),
                 subset(data_exp,treat=="unheated"))
model1_LFD_h<-lmer(LFD~1+(1|sire)+(1|dam)+(1|sire:dam),
                 subset(data_exp,treat=="heatmat"))
model1_MFD_uh<-lmer(MFD~1+(1|sire)+(1|dam)+(1|sire:dam),
                 subset(data_exp,treat=="unheated"))
model1_MFD_h<-lmer(MFD~1+(1|sire)+(1|dam)+(1|sire:dam),
                 subset(data_exp,treat=="heatmat"))
summary(model1_FFD_uh)
summary(model1_FFD_h)
summary(model1_LFD_uh)
summary(model1_LFD_h)
summary(model1_MFD_uh)
summary(model1_MFD_h)
```

### Proportions of variance

```{r}
Variance_FFD_uh<-as.data.frame(VarCorr(model1_FFD_uh))[,c(1,4)]
Variance_LFD_uh<-as.data.frame(VarCorr(model1_LFD_uh))[,c(1,4)]
Variance_MFD_uh<-as.data.frame(VarCorr(model1_MFD_uh))[,c(1,4)]
Variance_FFD_h<-as.data.frame(VarCorr(model1_FFD_h))[,c(1,4)]
Variance_LFD_h<-as.data.frame(VarCorr(model1_LFD_h))[,c(1,4)]
Variance_MFD_h<-as.data.frame(VarCorr(model1_MFD_h))[,c(1,4)]
# Intra-class correlations
PropVar_FFD_uh <- Variance_FFD_uh%>%mutate(propvar=vcov/sum(vcov))%>%
  mutate(variable="FFD",treat="unheated")
PropVar_LFD_uh <- Variance_LFD_uh%>%mutate(propvar=vcov/sum(vcov))%>%
  mutate(variable="LFD",treat="unheated")
PropVar_MFD_uh <- Variance_MFD_uh%>%mutate(propvar=vcov/sum(vcov))%>%
  mutate(variable="MFD",treat="unheated")
PropVar_FFD_h <- Variance_FFD_h%>%mutate(propvar=vcov/sum(vcov))%>%
  mutate(variable="FFD",treat="heated")
PropVar_LFD_h <- Variance_LFD_h%>%mutate(propvar=vcov/sum(vcov))%>%
  mutate(variable="LFD",treat="heated")
PropVar_MFD_h <- Variance_MFD_h%>%mutate(propvar=vcov/sum(vcov))%>%
  mutate(variable="MFD",treat="heated")
Props_var<-rbind(PropVar_FFD_uh,PropVar_LFD_uh,PropVar_MFD_uh,
                 PropVar_FFD_h,PropVar_LFD_h,PropVar_MFD_h)
```

```{r}
Props_var
```

### Heritability and maternal effects

```{r}
# h^2 (paternal effects)
her_FFD_uh<-as.numeric(4*subset(PropVar_FFD_uh,grp=="sire")[3])
her_FFD_h<-as.numeric(4*subset(PropVar_FFD_h,grp=="sire")[3])
her_LFD_uh<-as.numeric(4*subset(PropVar_LFD_uh,grp=="sire")[3])
her_LFD_h<-as.numeric(4*subset(PropVar_LFD_h,grp=="sire")[3])
her_MFD_uh<-as.numeric(4*subset(PropVar_MFD_uh,grp=="sire")[3])
her_MFD_h<-as.numeric(4*subset(PropVar_MFD_h,grp=="sire")[3])
# Because the additive genetic variance, VA,
# is expected to be four times the among pollen‐donor variance 
# (Falconer & Mackay, 1996; Lynch & Walsh, 1998)
her<-data.frame(value=rbind(her_FFD_uh,her_FFD_h,her_LFD_uh,
                       her_LFD_h,her_MFD_uh,her_MFD_h))%>%
  rownames_to_column()%>%
  mutate(variable=c("FFD","FFD","LFD","LFD","MFD","MFD"),
         treat=c("unheated","heated","unheated","heated",
                 "unheated","heated"),
         effect="Heritability")
```

```{r}
# maternal effects

# Maternal - paternal
# ----------------------
# summed effects

# Because the pollen‐recipient variance component contains a combination of
# genetic and environmental effects, we subtracted the additive genetic 
# (pollen donor) component from the pollen-recipient variance component
# before dividing the resulting estimate by VP to estimate 
# m2 (m2 = (Vpollen recipient − Vpollen donor)/VP).

mat_FFD_uh<-(Variance_FFD_uh[2,2]-Variance_FFD_uh[3,2])/
  as.numeric(Variance_FFD_uh%>%summarise(sum(vcov)))
mat_FFD_h<-(Variance_FFD_h[2,2]-Variance_FFD_h[3,2])/
  as.numeric(Variance_FFD_h%>%summarise(sum(vcov)))
mat_LFD_uh<-(Variance_LFD_uh[2,2]-Variance_LFD_uh[3,2])/
  as.numeric(Variance_LFD_uh%>%summarise(sum(vcov)))
mat_LFD_h<-(Variance_LFD_h[2,2]-Variance_LFD_h[3,2])/
  as.numeric(Variance_LFD_h%>%summarise(sum(vcov)))
mat_MFD_uh<-(Variance_MFD_uh[2,2]-Variance_MFD_uh[3,2])/
  as.numeric(Variance_MFD_uh%>%summarise(sum(vcov)))
mat_MFD_h<-(Variance_MFD_h[2,2]-Variance_MFD_h[3,2])/
  as.numeric(Variance_MFD_h%>%summarise(sum(vcov)))
mat<-data.frame(value=rbind(mat_FFD_uh,mat_FFD_h,mat_LFD_uh,
                       mat_LFD_h,mat_MFD_uh,mat_MFD_h))%>%
  rownames_to_column()%>%
  mutate(variable=c("FFD","FFD","LFD","LFD","MFD","MFD"),
         treat=c("unheated","heated","unheated","heated",
                 "unheated","heated"),
         effect="Maternal effects")
```

```{r}
her_mat<-rbind(her,mat)
her_mat
```

### Appendix S7: LRTs for variance components

```{r}
ranova(model1_FFD_uh)
ranova(model1_FFD_h)
ranova(model1_LFD_uh)
ranova(model1_LFD_h)
ranova(model1_MFD_uh)
ranova(model1_MFD_h)
```

## Additive genetic variation in flowering time plasticity

### Models

```{r}
model2_FFD<-lmer(FFD~treat+
                    (1|sire)+(1|dam)+
                    (1|treat:sire)+
                    #(1|treat:dam)+ # Removed to avoid singular fit
                    (1|sire:dam),
                  data_exp)
model2_LFD<-lmer(LFD~treat+
                    (1|sire)+(1|dam)+
                    #(1|treat:sire)+ # Removed to avoid singular fit
                    (1|treat:dam)+
                    (1|sire:dam),
                  data_exp)
model2_MFD<-lmer(MFD~treat+
                    (1|sire)+(1|dam)+
                    (1|treat:sire)+
                    #(1|treat:dam)+ # Removed to aovid singular fit
                    (1|sire:dam),
                  data_exp)
summary(model2_FFD)
summary(model2_LFD)
summary(model2_MFD)
```

### Proportions of variance

```{r}
# extract variance components
Variance_FFD_treat <- as.data.frame(VarCorr(model2_FFD))[,c(1,4)]  
Variance_LFD_treat <- as.data.frame(VarCorr(model2_LFD))[,c(1,4)]  
Variance_MFD_treat <- as.data.frame(VarCorr(model2_MFD))[,c(1,4)]  
# Intra-class correlation
PropVar_FFD_treat <- Variance_FFD_treat%>%
  mutate(propvar=vcov/sum(vcov))%>%mutate(variable="FFD")
PropVar_LFD_treat <- Variance_LFD_treat%>%
  mutate(propvar=vcov/sum(vcov))%>%mutate(variable="LFD")
PropVar_MFD_treat <- Variance_MFD_treat%>%
  mutate(propvar=vcov/sum(vcov))%>%mutate(variable="MFD")
# Proportional variance
Props_var_treat<-rbind(PropVar_FFD_treat,PropVar_LFD_treat,
                       PropVar_MFD_treat)
```

```{r}
Props_var_treat
```

### Heritability and maternal effects

```{r}
# h^2 (paternal effects)
her_FFD_treat<-4*subset(PropVar_FFD_treat,grp=="treat:sire")[3]
her_LFD_treat<-0
her_MFD_treat<-4*subset(PropVar_MFD_treat,grp=="treat:sire")[3]
# Because the additive genetic variance, VA,
# is expected to be four times the among pollen‐donor variance 
# (Falconer & Mackay, 1996; Lynch & Walsh, 1998)
her_treat<-data.frame(value=rbind(her_FFD_treat,
                                  her_LFD_treat,her_MFD_treat))%>%
  mutate(variable=c("FFD","LFD","MFD"),
         effect="Heritability")%>%
  rename(value=propvar)
```

```{r}
# maternal effects

# Maternal - paternal
# ----------------------
# summed effects

# Because the pollen‐recipient variance component contains a combination of
# genetic and environmental effects, we subtracted the additive genetic 
# (pollen donor) component from the pollen-recipient variance component
# before dividing the resulting estimate by VP to estimate 
# m2 (m2 = (Vpollen recipient − Vpollen donor)/VP).

mat_FFD_treat<-(0-
                  subset(Variance_FFD_treat,grp=="treat:sire")[2])/
  Variance_FFD_treat%>%summarise(sum(vcov))
mat_LFD_treat<-(subset(Variance_LFD_treat,grp=="treat:dam")[2]-
                  0)/
  Variance_LFD_treat%>%summarise(sum(vcov))
mat_MFD_treat<-(0-
                  subset(Variance_MFD_treat,grp=="treat:sire")[2])/
  Variance_MFD_treat%>%summarise(sum(vcov))
mat_treat<-data.frame(value=rbind(mat_FFD_treat,
                                  mat_LFD_treat,mat_MFD_treat))%>%
  mutate(variable=c("FFD","LFD","MFD"),
         effect="Maternal effects")%>%
  rename(value=vcov)
```

```{r}
her_mat_treat<-rbind(her_treat,mat_treat)
her_mat_treat
```

### Appendix S8: LRTs for variance components

```{r}
ranova(model2_FFD)
ranova(model2_LFD)
ranova(model2_MFD)
```

## Figure 1

```{r}
Props_var<-Props_var%>%
  mutate(variable=factor(variable,levels=c("FFD","MFD","LFD")),
         treat=factor(treat,levels=c("unheated","heated")),
         grp=factor(grp,levels=c("sire","dam","sire:dam","Residual")),
         sig=c("n","y","y","y",
               "n","y","n","Y",
               "n","y","y","y",
               "y","y","n","y",
               "n","y","n","y",
               "n","y","n","y"))%>% # Based on LRTs
  mutate(treat=fct_recode(treat,"control"="unheated"))
Props_var_treat<-Props_var_treat%>%
  mutate(variable=factor(variable,levels=c("FFD","MFD","LFD")),
         grp=factor(grp,levels=c("sire","dam","sire:dam",
                                 "treat:sire","treat:dam","Residual")),
         sig=c("n","y","y","n","y",
               "n","n","y","n","y",
               "n","y","y","n","y")) # Based on LRTs
```

```{r}
legend_fig1<-get_legend(
  ggplot(Props_var_treat%>%filter(grp!="Residual"),
         aes(x=variable,y=propvar,fill=grp,shape=sig))+
    geom_col(color="black",size=0.3)+
    geom_point(aes(x=variable,y=propvar,group=grp),
               position=position_stack(vjust=.5),show.legend=F,size=2)+
    my_theme_legend()+
    xlab(NULL)+ylab("Proportion of variance")+
    theme(legend.position="top",legend.title=element_blank())+
    scale_fill_discrete(labels=c("sire","dam","sire:dam",
                                 "treat:sire","treat:dam","residual"),
                        type=c("#00BFC4","#F8766D","#00BA38","#619CFF",
                               "#F564E3","grey"))+
    scale_shape_manual(values=c(NA,8))+ggtitle("B)")+
    theme(axis.text.y=element_blank(),axis.title.y=element_blank()))
fig1a<-ggplot(Props_var%>%filter(grp!="Residual"),
              aes(x=variable,y=propvar,fill=grp,shape=sig))+
  geom_col(color="black",size=0.3)+
  geom_point(aes(x=variable,y=propvar,group=grp),
             position=position_stack(vjust=.5),show.legend=F,size=2)+
  facet_wrap(~treat)+
  my_theme()+xlab(NULL)+ylab("Proportion of variance")+
  scale_fill_discrete(labels=c("sire","dam","sire:dam","residual"),
                      type=c("#00BFC4","#F8766D","#00BA38","grey"))+
  guides(fill=guide_legend(nrow=2))+
  theme(plot.title = element_text(hjust =-0.05,vjust=0))+
  scale_shape_manual(values=c(NA,8))+ggtitle("A)")
fig1b<-ggplot(Props_var_treat%>%filter(grp!="Residual"),
              aes(x=variable,y=propvar,fill=grp,shape=sig))+
  geom_col(color="black",size=0.3)+
  geom_point(aes(x=variable,y=propvar,group=grp),
             position=position_stack(vjust=.5),show.legend=F,size=2)+
  my_theme()+
  xlab(NULL)+ylab("Proportion of variance")+
  scale_fill_discrete(labels=c("sire","dam","sire:dam",
                               "treat:sire","treat:dam","residual"),
                      type=c("#00BFC4","#F8766D","#00BA38","#619CFF",
                             "#F564E3","grey"))+
  theme(plot.title = element_text(hjust =-0.05,vjust=4))+
  scale_shape_manual(values=c(NA,8))+ggtitle("B)")+
  theme(axis.title.y=element_blank())
fig1_without_legend<-cowplot::plot_grid(fig1a,fig1b,
                                        ncol=2,align="hv",axis="t",
                                        rel_widths = c(1.8, 1))
fig1<-grid.arrange(legend_fig1,fig1_without_legend,ncol=1,
                   heights=c(0.05,0.95))
ggsave(file="output/figures/fig1.tiff", plot=fig1,width=10,height=5,dpi=300,
       compression="lzw")
```

## Table 1

```{r}
her_mat<-her_mat%>%
  mutate(variable=factor(variable,levels=c("FFD","MFD","LFD")),
         treat=factor(treat,levels=c("unheated","heated")))%>%
  dplyr::select(-rowname)%>%
  pivot_wider(names_from="effect",values_from="value")
her_mat
her_mat_treat<-her_mat_treat%>%
  pivot_wider(names_from="effect",values_from="value")
her_mat_treat
```

## Figure 2

Main effect of temperature treatment significant for FFD, LFD and MFD.

```{r}
data_fig2<-rbind(
  data.frame(ggpredict(model2_FFD,terms="treat"))%>%mutate(variable="FFD"),
  data.frame(ggpredict(model2_MFD,terms="treat"))%>%mutate(variable="MFD"),
  data.frame(ggpredict(model2_LFD,terms="treat"))%>%mutate(variable="LFD"),
  data.frame(ggpredict(model2_FFD,terms=c("treat","sire"),type="random"))%>%
    mutate(variable="FFD"),
  data.frame(ggpredict(model2_MFD,terms=c("treat","sire"),type="random"))%>%
    mutate(variable="MFD"))%>%
  mutate(variable=factor(variable,levels=c("FFD","MFD","LFD")),
         x=factor(x,levels=c("unheated","heatmat"),
                  labels=c("unheated","heated")),
         type=ifelse(group==1,"mean","individual"))
```

```{r}
fig2<-ggplot()+
  # Mean responses
  geom_point(data=data_fig2%>%filter(group==1)%>%
               mutate(x=factor(x,levels=c("unheated","heated")))%>%
               mutate(x=fct_recode(x,"control"="unheated")),
             aes(x=x,y=predicted,color=variable),size=2)+
  geom_errorbar(data=data_fig2%>%filter(group==1)%>%
                  mutate(x=factor(x,levels=c("unheated","heated")))%>%
                  mutate(x=fct_recode(x,"control"="unheated")),
                aes(x=x,y=predicted,ymin=conf.low,ymax=conf.high,
                    color=variable),
                width=0.05,size=0.5)+
  geom_line(data=data_fig2%>%filter(group==1)%>%
              mutate(x=factor(x,levels=c("unheated","heated")))%>%
              mutate(x=fct_recode(x,"control"="unheated")),
            aes(x=x,y=predicted,group=variable,color=variable),size=0.5)+
  # Response of each sire for FFD
  geom_line(data=data_fig2%>%filter(group!=1)%>%
              mutate(x=factor(x,levels=c("unheated","heated")))%>%
              mutate(x=fct_recode(x,"control"="unheated"))%>%
              filter(variable=="FFD"),
            aes(x=x,y=predicted,group=group),
            linewidth=0.2,color="#F8766D",alpha=0.5)+
  # Response of each sire for MFD
  geom_line(data=data_fig2%>%filter(group!=1)%>%
              mutate(x=factor(x,levels=c("unheated","heated")))%>%
              mutate(x=fct_recode(x,"control"="unheated"))%>%
              filter(variable=="MFD"),
            aes(x=x,y=predicted,group=group),
            linewidth=0.2,color="#00BA38",alpha=0.5)+
  my_theme_legend()+theme(legend.title=element_blank())+
  theme(plot.title = element_text(hjust =-0.2,vjust=-1.5))+
  theme(legend.position = c(0.84, 0.88))+
  theme(legend.background = element_rect(fill = "transparent"))+
  xlab(NULL)+ylab("Flowering date")+scale_x_discrete(expand=c(0.1,0.1))
ggsave(file="output/figures/fig2.tiff", plot=fig2,width=5,height=4,dpi=300,
       compression="lzw")
```

## Hypothesis 1

Temperature-dependent selection has resulted in small-scale genetic differentiation, and additive genetic variation in phenology is correlated with soil temperature in a counter-gradient fashion. At a given temperature, individuals with parents from warmer soils are predicted to start vegetative growth and flower later than plants with parents from colder soils.

### Models temperatures sire and dam

```{r}
model3_FFD_uh<-lmer(FFD~temp_sire+temp_dam+
                      (1|sire)+(1|dam)+(1|sire:dam),
                 subset(data_exp,treat=="unheated"&!is.na(FFD)))
model3_FFD_h<-lmer(FFD~temp_sire+temp_dam+
                     (1|sire)+(1|dam)+(1|sire:dam),
                 subset(data_exp,treat=="heatmat"&!is.na(FFD)))
model3_LFD_uh<-lmer(LFD~temp_sire+temp_dam+
                      (1|sire)+(1|dam)+(1|sire:dam),
                 subset(data_exp,treat=="unheated"&!is.na(LFD)))
model3_LFD_h<-lmer(LFD~temp_sire+temp_dam+
                     (1|sire)+(1|dam)+(1|sire:dam),
                 subset(data_exp,treat=="heatmat"&!is.na(LFD)))
model3_MFD_uh<-lmer(MFD~temp_sire+temp_dam+
                      (1|sire)+(1|dam)+(1|sire:dam),
                 subset(data_exp,treat=="unheated"&!is.na(MFD)))
model3_MFD_h<-lmer(MFD~temp_sire+temp_dam+
                     #(1|sire) # Removed to avoid singular fit
                     (1|dam)+(1|sire:dam),
                 subset(data_exp,treat=="heatmat"&!is.na(MFD)))
summary(model3_FFD_uh)
summary(model3_FFD_h)
summary(model3_LFD_uh)
summary(model3_LFD_h)
summary(model3_MFD_uh)
summary(model3_MFD_h)
```

#### Prediction results

```{r}
ggpredict(model3_MFD_h,terms=c("temp_sire[5]"))
ggpredict(model3_MFD_h,terms=c("temp_sire[45.5]"))
```

#### Table 2

```{r}
tab_model(model3_FFD_uh,model3_MFD_uh,model3_LFD_uh,
          transform=NULL,show.intercept=F,show.ci=F,show.se=T,show.stat=T,
          show.r2=T,show.icc=F,show.re.var=F,show.ngroups=F,digits=3,
          dv.labels=c("FFD","MFD","LFD"),
          file="output/tables/Table2_uh.html")
tab_model(model3_FFD_h,model3_MFD_h,model3_LFD_h,
          transform=NULL,show.intercept=F,show.ci=F,show.se=T,show.stat=T,
          show.r2=T,show.icc=F,show.re.var=F,show.ngroups=F,digits=3,
          dv.labels=c("FFD","MFD","LFD"),
          file="output/tables/Table2_h.html")
```

#### Proportions of variance

Proportion of variance explained by temperature of origin of sires and dams = proportion of variance explained by fixed effects: marginal R2.

```{r}
PropVar_FFD_uh_temp<-as.numeric(r2_nakagawa(model3_FFD_uh)$R2_marginal)
PropVar_FFD_h_temp<-as.numeric(r2_nakagawa(model3_FFD_h)$R2_marginal)
PropVar_LFD_uh_temp<-as.numeric(r2_nakagawa(model3_LFD_uh)$R2_marginal)
PropVar_LFD_h_temp<-as.numeric(r2_nakagawa(model3_LFD_h)$R2_marginal)
PropVar_MFD_uh_temp<-as.numeric(r2_nakagawa(model3_MFD_uh)$R2_marginal)
PropVar_MFD_h_temp<-as.numeric(r2_nakagawa(model3_MFD_h)$R2_marginal)
Props_var_temp<-data.frame(
  values=c(PropVar_FFD_uh_temp,PropVar_FFD_h_temp,
           PropVar_LFD_uh_temp,PropVar_LFD_h_temp,
           PropVar_MFD_uh_temp,PropVar_MFD_h_temp),
  variable=c("FFD","FFD","LFD","LFD","MFD","MFD"),
  treat=c("unheated","heated","unheated","heated",
                 "unheated","heated"))
```

```{r}
Props_var_temp
```

#### Temperature dam and sire: partR2

```{r eval=FALSE, include=FALSE}
partR2_model3_FFD_uh<-partR2(model3_FFD_uh,
                             partvars=c("temp_sire","temp_dam"),
                             R2_type="marginal",nboot=1000)
partR2_model3_FFD_h<-partR2(model3_FFD_h,
                             partvars=c("temp_sire","temp_dam"),
                             R2_type="marginal",nboot=1000)
partR2_model3_LFD_uh<-partR2(model3_LFD_uh,
                             partvars=c("temp_sire","temp_dam"),
                             R2_type="marginal",nboot=1000)
partR2_model3_LFD_h<-partR2(model3_LFD_h,
                             partvars=c("temp_sire","temp_dam"),
                             R2_type="marginal",nboot=1000)
partR2_model3_MFD_uh<-partR2(model3_MFD_uh,
                             partvars=c("temp_sire","temp_dam"),
                             R2_type="marginal",nboot=1000)
partR2_model3_MFD_h<-partR2(model3_MFD_h,
                             partvars=c("temp_sire","temp_dam"),
                             R2_type="marginal",nboot=1000)
```

```{r eval=FALSE, include=FALSE}
save(partR2_model3_FFD_uh, file = "output/models/partR2_model3_FFD_uh.rda")
save(partR2_model3_FFD_h, file = "output/models/partR2_model3_FFD_h.rda")
save(partR2_model3_LFD_uh, file = "output/models/partR2_model3_LFD_uh.rda")
save(partR2_model3_LFD_h, file = "output/models/partR2_model3_LFD_h.rda")
save(partR2_model3_MFD_uh, file = "output/models/partR2_model3_MFD_uh.rda")
save(partR2_model3_MFD_h, file = "output/models/partR2_model3_MFD_h.rda")
```

```{r}
Props_var_temp_mf<-
  rbind(
    partR2_model3_FFD_uh$R2%>%mutate(variable="FFD",treat="unheated"),
    partR2_model3_FFD_h$R2%>%mutate(variable="FFD",treat="heated"),
    partR2_model3_LFD_uh$R2%>%mutate(variable="LFD",treat="unheated"),
    partR2_model3_LFD_h$R2%>%mutate(variable="LFD",treat="heated"),
    partR2_model3_MFD_uh$R2%>%mutate(variable="MFD",treat="unheated"),
    partR2_model3_MFD_h$R2%>%mutate(variable="MFD",treat="heated")
    )
```

```{r}
Props_var_temp_mf
```

#### Figure 3

```{r}
fig3<-ggplot()+
  geom_bar(data=Props_var_temp_mf%>%
             filter(term!="temp_sire+temp_dam"&term!="Full")%>%
             mutate(term=factor(term,levels=c("temp_sire","temp_dam"),
                                labels=c("Sire origin temperature",
                                           "Dam origin temperature")),
                    variable=factor(variable,levels=c("veg","FFD","MFD","LFD")),
                    sig=c("n","n","n","n",
                          "n","n","n","n","n","n","y","n"))%>%
               mutate(treat=fct_recode(treat,"control"="unheated")),
           aes(x=variable,y=estimate,fill=term),
           stat="identity",color="black",size=0.2)+
  facet_wrap(~factor(treat,levels=c("control","heated")))+
  geom_point(data=Props_var_temp_mf%>%
               filter(term!="temp_sire+temp_dam"&term!="Full")%>%
               mutate(term=factor(term,levels=c("temp_sire","temp_dam"),
                                  labels=c("Sire origin temperature",
                                           "Dam origin temperature")),
                      variable=factor(variable,
                                      levels=c("veg","FFD","MFD","LFD")),
                      sig=c("n","n","n","n",
                            "n","n","n","n","n","n","y","n"))%>%
               mutate(treat=fct_recode(treat,"control"="unheated")),
             aes(x=variable,y=estimate,group=term,shape=sig),
             position=position_stack(vjust=.5),show.legend=F,size=2)+
  my_theme_legend()+scale_fill_discrete(type=c("#00BFC4","#F8766D"))+
  scale_shape_manual(values=c(NA,8))+
  xlab(NULL)+ylab("Proportion of variance")+
  theme(legend.position="top",legend.title=element_blank())
fig3
ggsave(file="output/figures/fig3.tiff", plot=fig3,width=5,height=5,dpi=300,
       compression="lzw")
```

#### Appendix S9

```{r}
AppS9<-ggplot(data=Props_var_temp_mf%>%
             filter(term!="temp_sire+temp_dam"&term!="Full")%>%
             mutate(term=factor(term,levels=c("temp_sire","temp_dam"),
                                labels=c("Sire origin temperature",
                                           "Dam origin temperature")),
                    variable=factor(variable,levels=c("FFD","MFD","LFD")),
                    sig=c("n","n","n","n",
                          "n","n","n","n","n","n","y","n"))%>%
               mutate(treat=fct_recode(treat,"control"="unheated")),
           aes(x=variable,y=estimate,ymin=CI_lower,ymax=CI_upper,color=term))+
  geom_point(size=2,position=position_dodge(0.4),width=0.2)+
  geom_errorbar(position=position_dodge(0.4),width=0.2)+
  facet_wrap(~factor(treat,levels=c("control","heated")))+
  my_theme_legend()+
  scale_color_discrete(type=c("#00BFC4","#F8766D"))+
  xlab(NULL)+ylab("Proportion of variance")+
  theme(legend.position="top",legend.title=element_blank())
AppS9
ggsave(file="output/figures/AppS9.tiff", plot=AppS9,width=7,height=8,dpi=300,
       compression="lzw")
```

### Models mid-parental values

```{r}
data_exp<-data_exp%>%
  mutate(temp_midP=(temp_sire+temp_dam)/2)
```

```{r}
model3_FFD_uh_midP<-lmer(FFD~temp_midP+
                      (1|sire)+(1|dam)+(1|sire:dam),
                 subset(data_exp,treat=="unheated"&!is.na(FFD)))
model3_FFD_h_midP<-lmer(FFD~temp_midP+
                     (1|sire)+(1|dam)+(1|sire:dam),
                 subset(data_exp,treat=="heatmat"&!is.na(FFD)))
model3_LFD_uh_midP<-lmer(LFD~temp_midP+
                      (1|sire)+(1|dam)+(1|sire:dam),
                 subset(data_exp,treat=="unheated"&!is.na(LFD)))
model3_LFD_h_midP<-lmer(LFD~temp_midP+
                     (1|sire)+(1|dam)+(1|sire:dam),
                 subset(data_exp,treat=="heatmat"&!is.na(LFD)))
model3_MFD_uh_midP<-lmer(MFD~temp_midP+
                      (1|sire)+(1|dam)+(1|sire:dam),
                 subset(data_exp,treat=="unheated"&!is.na(MFD)))
model3_MFD_h_midP<-lmer(MFD~temp_midP+
                     (1|sire)+(1|dam)+(1|sire:dam),
                 subset(data_exp,treat=="heatmat"&!is.na(MFD)))
summary(model3_FFD_uh_midP)
summary(model3_FFD_h_midP)
summary(model3_LFD_uh_midP)
summary(model3_LFD_h_midP)
summary(model3_MFD_uh_midP)
summary(model3_MFD_h_midP)
```

#### Appendix S4

```{r}
tab_model(model3_FFD_uh_midP,model3_MFD_uh_midP,
          model3_LFD_uh_midP,
          transform=NULL,show.intercept=F,show.ci=F,show.se=T,show.stat=T,
          show.r2=T,show.icc=F,show.re.var=F,show.ngroups=F,digits=3,
          dv.labels=c("FFD","MFD","LFD"),
          file="output/tables/AppS4_uh.html")
tab_model(model3_FFD_h_midP,model3_MFD_h_midP,
          model3_LFD_h_midP,
          transform=NULL,show.intercept=F,show.ci=F,show.se=T,show.stat=T,
          show.r2=T,show.icc=F,show.re.var=F,show.ngroups=F,digits=3,
          dv.labels=c("FFD","MFD","LFD"),
          file="output/tables/AppS4_h.html")
```

### Models breeding values (Appendix S6)

Get breeding values/BLUPs from models in Prediction 1.

```{r}
averages_resp<-rbind(
  data.frame(average=as.numeric(fixef(model1_FFD_uh)))%>%
    mutate(name="FFD_uh_resp"),
  data.frame(average=as.numeric(fixef(model1_FFD_h)))%>%
    mutate(name="FFD_h_resp"),
  data.frame(average=as.numeric(fixef(model1_MFD_uh)))%>%
    mutate(name="MFD_uh_resp"),
  data.frame(average=as.numeric(fixef(model1_MFD_h)))%>%
    mutate(name="MFD_h_resp"),
  data.frame(average=as.numeric(fixef(model1_LFD_uh)))%>%
    mutate(name="LFD_uh_resp"),
  data.frame(average=as.numeric(fixef(model1_LFD_h)))%>%
    mutate(name="LFD_h_resp")
  )%>%
  separate(name,into=c("var","treat","type"))
```

```{r}
breedV_resp_sires<-left_join(
  left_join(
    left_join(
      left_join(
        left_join(
            ranef(model1_FFD_uh)$sire%>%rownames_to_column(var="sire")%>%
              rename(breedV_FFD_uh_resp_sire="(Intercept)"),
          ranef(model1_FFD_h)$sire%>%rownames_to_column(var="sire")%>%
            rename(breedV_FFD_h_resp_sire="(Intercept)")
          ),
        ranef(model1_MFD_uh)$sire%>%rownames_to_column(var="sire")%>%
          rename(breedV_MFD_uh_resp_sire="(Intercept)")
        ),
      ranef(model1_MFD_h)$sire%>%rownames_to_column(var="sire")%>%
        rename(breedV_MFD_h_resp_sire="(Intercept)")
      ),
    ranef(model1_LFD_uh)$sire%>%rownames_to_column(var="sire")%>%
      rename(breedV_LFD_uh_resp_sire="(Intercept)")
    ),
  ranef(model1_LFD_h)$sire%>%rownames_to_column(var="sire")%>%
    rename(breedV_LFD_h_resp_sire="(Intercept)")
  )%>%
  pivot_longer(cols=breedV_FFD_uh_resp_sire:breedV_LFD_h_resp_sire,
               names_to=c("remove","var","treat","type","comp"),names_sep="_",
               values_to="breedV")%>%
  dplyr::select(-remove)%>%
  left_join(averages_resp)%>%mutate(breedV_av=breedV+average)
```

```{r}
breedV_resp_dams<-left_join(
  left_join(
    left_join(
      left_join(
        left_join(
            ranef(model1_FFD_uh)$dam%>%rownames_to_column(var="dam")%>%
              rename(breedV_FFD_uh_resp_dam="(Intercept)"),
          ranef(model1_FFD_h)$dam%>%rownames_to_column(var="dam")%>%
            rename(breedV_FFD_h_resp_dam="(Intercept)")
          ),
        ranef(model1_MFD_uh)$dam%>%rownames_to_column(var="dam")%>%
          rename(breedV_MFD_uh_resp_dam="(Intercept)")
        ),
      ranef(model1_MFD_h)$dam%>%rownames_to_column(var="dam")%>%
        rename(breedV_MFD_h_resp_dam="(Intercept)")
      ),
    ranef(model1_LFD_uh)$dam%>%rownames_to_column(var="dam")%>%
      rename(breedV_LFD_uh_resp_dam="(Intercept)")
    ),
  ranef(model1_LFD_h)$dam%>%rownames_to_column(var="dam")%>%
    rename(breedV_LFD_h_resp_dam="(Intercept)")
  )%>%
  pivot_longer(cols=breedV_FFD_uh_resp_dam:breedV_LFD_h_resp_dam,
               names_to=c("remove","var","treat","type","comp"),names_sep="_",
               values_to="breedV")%>%
  dplyr::select(-remove)%>%
  left_join(averages_resp)%>%mutate(breedV_av=breedV+average)
```

Add origin temperatures:

```{r}
breedV_resp_sires<-breedV_resp_sires%>%
  left_join(data_exp%>%dplyr::select(sire,temp_sire)%>%unique())
breedV_resp_dams<-breedV_resp_dams%>%
  left_join(data_exp%>%dplyr::select(dam,temp_dam)%>%unique())
```

#### Sire breeding values ~ origin temperatures

```{r}
mod_bV_resp_temp_sire_FFD_uh<-lm(breedV~temp_sire,
                                 subset(breedV_resp_sires,
                                        var=="FFD"&treat=="uh"))
mod_bV_resp_temp_sire_FFD_h<-lm(breedV~temp_sire,
                                 subset(breedV_resp_sires,
                                        var=="FFD"&treat=="h"))
mod_bV_resp_temp_sire_MFD_uh<-lm(breedV~temp_sire,
                                 subset(breedV_resp_sires,
                                        var=="MFD"&treat=="uh"))
mod_bV_resp_temp_sire_MFD_h<-lm(breedV~temp_sire,
                                 subset(breedV_resp_sires,
                                        var=="MFD"&treat=="h"))
mod_bV_resp_temp_sire_LFD_uh<-lm(breedV~temp_sire,
                                 subset(breedV_resp_sires,
                                        var=="LFD"&treat=="uh"))
mod_bV_resp_temp_sire_LFD_h<-lm(breedV~temp_sire,
                                 subset(breedV_resp_sires,
                                        var=="LFD"&treat=="h"))
```

```{r}
summary(mod_bV_resp_temp_sire_FFD_uh)
summary(mod_bV_resp_temp_sire_FFD_h)
summary(mod_bV_resp_temp_sire_MFD_uh)
summary(mod_bV_resp_temp_sire_MFD_h) #*
summary(mod_bV_resp_temp_sire_LFD_uh)
summary(mod_bV_resp_temp_sire_LFD_h)
```

##### Figure S6A

```{r}
figS6A<-plot(ggpredict(mod_bV_resp_temp_sire_MFD_h,terms="temp_sire"),
             add.data=T)+
  xlab("Sire origin temperature (ºC)")+
  ylab("Sire breeding value for the\nresponse of median flowering date")+
  my_theme()+ggtitle(NULL)
figS6A
ggsave(file="output/figures/figS6A.tiff", plot=figS6A,width=5,height=4,dpi=300,
       compression="lzw")
```

#### Dam breeding values ~ origin temperatures

```{r}
mod_bV_resp_temp_dam_FFD_uh<-lm(breedV~temp_dam,
                                 subset(breedV_resp_dams,
                                        var=="FFD"&treat=="uh"))
mod_bV_resp_temp_dam_FFD_h<-lm(breedV~temp_dam,
                                 subset(breedV_resp_dams,
                                        var=="FFD"&treat=="h"))
mod_bV_resp_temp_dam_MFD_uh<-lm(breedV~temp_dam,
                                 subset(breedV_resp_dams,
                                        var=="MFD"&treat=="uh"))
mod_bV_resp_temp_dam_MFD_h<-lm(breedV~temp_dam,
                                 subset(breedV_resp_dams,
                                        var=="MFD"&treat=="h"))
mod_bV_resp_temp_dam_LFD_uh<-lm(breedV~temp_dam,
                                 subset(breedV_resp_dams,
                                        var=="LFD"&treat=="uh"))
mod_bV_resp_temp_dam_LFD_h<-lm(breedV~temp_dam,
                                 subset(breedV_resp_dams,
                                        var=="LFD"&treat=="h"))
```

```{r}
summary(mod_bV_resp_temp_dam_FFD_uh)
summary(mod_bV_resp_temp_dam_FFD_h)
summary(mod_bV_resp_temp_dam_MFD_uh)
summary(mod_bV_resp_temp_dam_MFD_h)
summary(mod_bV_resp_temp_dam_LFD_uh)
summary(mod_bV_resp_temp_dam_LFD_h)
# All NS
```

## Hypothesis 2

Genetic differentiation related to soil temperature involves also plasticity of development rate. Differences in phenology between high and low temperatures are predicted to be smaller for plants with parents from warm soils than for plants from colder soils.

### Models temperatures sire and dam

```{r}
model4_FFD<-lmer(FFD~treat*(temp_sire+temp_dam)+
                    (1|sire)+(1|dam)+
                    (1|sire:dam),
                  data_exp)
model4_LFD<-lmer(LFD~treat*(temp_sire+temp_dam)+
                    (1|sire)+(1|dam)+
                    (1|sire:dam),
                  data_exp)
model4_MFD<-lmer(MFD~treat*(temp_sire+temp_dam)+
                    (1|sire)+(1|dam)+
                    (1|sire:dam),
                  subset(data_exp,!is.na(MFD)))
summary(model4_FFD)
summary(model4_LFD)
summary(model4_MFD)
```

#### Prediction results

```{r}
ggpredict(model4_MFD,terms=c("temp_sire[5]"))
ggpredict(model4_MFD,terms=c("temp_sire[45.5]"))
```

#### Table 3

```{r}
tab_model(model4_FFD,model4_MFD,model4_LFD,
          transform=NULL,show.intercept=F,show.ci=F,show.se=T,show.stat=T,
          show.r2=T,show.icc=F,show.re.var=F,show.ngroups=F,digits=3,
          dv.labels=c("FFD","MFD","LFD"),
          file="output/tables/Table3.html")
```

#### Figure 4

```{r}
fig4<-ggplot()+
  geom_ribbon(data=data.frame(ggpredict(model3_MFD_h,
                                        terms=c("temp_sire"))),
              aes(x=x,y=predicted,ymin=conf.low,ymax=conf.high),
              fill="red",alpha=0.25)+
  geom_line(data=data.frame(ggpredict(model3_MFD_uh,
                                      terms=c("temp_sire"))),
            aes(x=x,y=predicted),color="blue",linetype="dashed")+
  geom_ribbon(data=data.frame(ggpredict(model3_MFD_uh,
                                        terms=c("temp_sire"))),
              aes(x=x,y=predicted,ymin=conf.low,ymax=conf.high),
              fill="blue",alpha=0.25)+
  geom_line(data=data.frame(ggpredict(model3_MFD_h,
                                      terms=c("temp_sire"))),
            aes(x=x,y=predicted),color="red")+
  geom_ribbon(data=data.frame(ggemmeans(model4_MFD,
                                        terms=c("temp_sire"))),
              aes(x=x,y=predicted,ymin=conf.low,ymax=conf.high),
              fill="darkgrey",alpha=0.25)+
  geom_line(data=data.frame(ggemmeans(model4_MFD,
                                      terms=c("temp_sire"))),
            aes(x=x,y=predicted),color="black",size=0.75)+
  geom_jitter(data=data_exp%>%
                mutate(treat=fct_recode(treat,"control"="unheated")),
              aes(x=temp_sire,y=MFD,color=treat),size=1,alpha=0.1)+
  scale_color_manual(values=c("blue","red"),limits=c("control","heatmat"),
                     labels=c("control","heated"))+
  my_theme_legend()+xlab("Sire origin temperature (ºC)")+
  ylab("Median flowering date")+
  theme(legend.position="top",legend.title=element_blank())+
  theme(legend.position = c(0.25, 0.95),legend.direction="horizontal")+
  theme(legend.background = element_rect(fill = "transparent"),
        legend.spacing.x = unit(0.1, 'cm'))
fig4
ggsave(file="output/figures/fig4.tiff", plot=fig4,width=4,height=4,dpi=300,
       compression="lzw")
```

### Models mid-parental values

```{r}
model4_FFD_midP<-lmer(FFD~treat*temp_midP+
                    (1|sire)+(1|dam)+
                    (1|sire:dam),
                  data_exp)
model4_LFD_midP<-lmer(LFD~treat*temp_midP+
                    (1|sire)+(1|dam)+
                    (1|sire:dam),
                  data_exp)
model4_MFD_midP<-lmer(MFD~treat*temp_midP+
                    (1|sire)+(1|dam)+
                    (1|sire:dam),
                  data_exp)
summary(model4_FFD_midP)
summary(model4_LFD_midP)
summary(model4_MFD_midP)
```

#### Appendix S5

```{r}
tab_model(model4_FFD_midP,model4_MFD_midP,model4_LFD_midP,
          transform=NULL,show.intercept=F,show.ci=F,show.se=T,show.stat=T,
          show.r2=T,show.icc=F,show.re.var=F,show.ngroups=F,digits=3,
          dv.labels=c("FFD","MFD","LFD"),
          file="output/tables/AppS5.html")
```

### Models breeding values (Appendix S6)

Get breeding values/BLUPs from models in Prediction 2.

```{r}
averages_plast<-rbind(
  data.frame(ggpredict(model2_FFD))[1:2]%>%
    pivot_wider(names_from=treat.x,values_from=treat.predicted)%>%
    mutate(slope=heatmat-unheated,name="FFD_plast"),
  data.frame(ggpredict(model2_MFD))[1:2]%>%
    pivot_wider(names_from=treat.x,values_from=treat.predicted)%>%
    mutate(slope=heatmat-unheated,name="MFD_plast"),
  data.frame(ggpredict(model2_LFD))[1:2]%>%
    pivot_wider(names_from=treat.x,values_from=treat.predicted)%>%
    mutate(slope=heatmat-unheated,name="LFD_plast")
  )%>%
  separate(name,into=c("var","type"))%>%
  rename(average=slope,av_heatmat=heatmat,av_unheated=unheated)%>%
  dplyr::select(var,type,average,av_heatmat,av_unheated)
```

```{r}
breedV_plast_sires<-left_join(
  ranef(model2_FFD)$`treat:sire`%>%
    rownames_to_column(var="treat_sire")%>%
    rename(breedV_FFD_plast_sire="(Intercept)")%>%
    separate(treat_sire,into=c("treat","sire"))%>%
    pivot_wider(names_from="treat",
                values_from="breedV_FFD_plast_sire",
                names_prefix="breedV_FFD_plast_sire_"),
  ranef(model2_MFD)$`treat:sire`%>%
    rownames_to_column(var="treat_sire")%>%
    rename(breedV_MFD_plast_sire="(Intercept)")%>%
    separate(treat_sire,into=c("treat","sire"))%>%
    pivot_wider(names_from="treat",
                values_from="breedV_MFD_plast_sire",
                names_prefix="breedV_MFD_plast_sire_")
  )%>%
  pivot_longer(cols=breedV_FFD_plast_sire_heatmat:breedV_MFD_plast_sire_unheated,
               names_to=c("remove","var","type","comp","treat"),names_sep="_",
               values_to="breedV")%>%
  dplyr::select(-remove)%>%
  pivot_wider(names_from="treat",values_from="breedV",names_prefix="breedV_")%>%
  mutate(breedV=breedV_heatmat-breedV_unheated)%>% # Difference heated-unheated
  left_join(averages_plast)%>%mutate(breedV_av=average+breedV)
```

```{r}
breedV_plast_dams<-
  ranef(model2_LFD)$`treat:dam`%>%
    rownames_to_column(var="treat_dam")%>%
    rename(breedV_LFD_plast_dam="(Intercept)")%>%
    separate(treat_dam,into=c("treat","dam"))%>%
    pivot_wider(names_from="treat",
                values_from="breedV_LFD_plast_dam",
                names_prefix="breedV_LFD_plast_dam_")%>%
    pivot_longer(cols=breedV_LFD_plast_dam_heatmat:breedV_LFD_plast_dam_unheated,
               names_to=c("remove","var","type","comp","treat"),names_sep="_",
               values_to="breedV")%>%
  dplyr::select(-remove)%>%
  pivot_wider(names_from="treat",values_from="breedV",names_prefix="breedV_")%>%
  mutate(breedV=breedV_heatmat-breedV_unheated)%>% # Difference heated-unheated
  left_join(averages_plast)%>%mutate(breedV_av=average+breedV)
```

Add origin temperatures:

```{r}
breedV_plast_sires<-breedV_plast_sires%>%
  left_join(data_exp%>%dplyr::select(sire,temp_sire)%>%unique())
breedV_plast_dams<-breedV_plast_dams%>%
  left_join(data_exp%>%dplyr::select(dam,temp_dam)%>%unique())
```

#### Sire breeding values ~ origin temperatures

```{r}
mod_bV_plast_temp_sire_FFD<-lm(breedV~temp_sire,
                                 subset(breedV_plast_sires,var=="FFD"))
mod_bV_plast_temp_sire_MFD<-lm(breedV~temp_sire,
                                 subset(breedV_plast_sires,var=="MFD"))
```

```{r}
summary(mod_bV_plast_temp_sire_FFD)
summary(mod_bV_plast_temp_sire_MFD)
# All NS
```

#### Dam breeding values ~ origin temperatures

```{r}
mod_bV_plast_temp_dam_LFD<-lm(breedV~temp_dam,
                                 subset(breedV_plast_dams,var=="LFD"))
```

```{r}
summary(mod_bV_plast_temp_dam_LFD)
# All NS
```

# New appendix: relations with flower number

```{r}
data_exp<-data_exp%>%mutate(total_n_fl_log=log(total_n_fl))
```

```{r}
model_FFD_fln<-lm(FFD~total_n_fl_log,data_exp%>%filter(total_n_fl>0))
model_MFD_fln<-lm(MFD~total_n_fl_log,data_exp%>%filter(total_n_fl>0))
model_LFD_fln<-lm(LFD~total_n_fl_log,data_exp%>%filter(total_n_fl>0))
```

```{r}
summary(model_FFD_fln)
summary(model_MFD_fln)
summary(model_LFD_fln)
```

```{r}
data_plot_fln<-rbind(
  data.frame(ggpredict(model_FFD_fln))%>%mutate(variable="FFD"),
  data.frame(ggpredict(model_MFD_fln))%>%mutate(variable="MFD"),
  data.frame(ggpredict(model_LFD_fln))%>%mutate(variable="LFD")
)%>%
  mutate(variable=factor(variable,levels = c("FFD", "MFD", "LFD")))
```

```{r}
r2_values <- data.frame(
  variable = c("FFD", "MFD", "LFD"),
  r2 = c(round(summary(model_FFD_fln)$r.squared,3),
         round(summary(model_MFD_fln)$r.squared,3),
         round(summary(model_LFD_fln)$r.squared,3))
)
```

```{r}
data_plot_fln <- data_plot_fln %>%
  left_join(r2_values, by = "variable")
```

```{r}
custom_labeller <- as_labeller(
  setNames(
    paste0(c("FFD", "MFD", "LFD"), "\nR² = ", c(0.362, 0.040, 0.107)), 
    c("FFD", "MFD", "LFD")
    )
  )
```


```{r}
ggplot()+
  geom_point(data=data_exp%>%select(FFD,LFD,MFD,total_n_fl_log)%>%
               pivot_longer(cols=c(FFD, MFD, LFD),names_to="variable",
                            values_to = "value")%>%
               mutate(variable=factor(variable,levels=c("FFD","MFD","LFD"))),
             aes(x=total_n_fl_log,y=value),
             alpha=0.25,size=1.5,show.legend=F)+
  facet_wrap(~variable, labeller = custom_labeller) + 
  geom_ribbon(data=data_plot_fln,
              aes(x=total_n_fl_log.x,y=total_n_fl_log.predicted,
                  ymin=total_n_fl_log.conf.low,ymax=total_n_fl_log.conf.high),
              alpha=0.25)+
  geom_line(data=data_plot_fln,
            aes(x=total_n_fl_log.x,y=total_n_fl_log.predicted))+
  my_theme()+theme(legend.title=element_blank())+
  xlab("Total number of flowers (log)")+ylab("Flowering date")
ggsave(file="output/figures/new_appendix.tiff", width=12,height=5,dpi=300,
       compression="lzw")
```

Correlations:

```{r}
with(data_exp,cor(FFD,total_n_fl,use="complete.obs"))
with(data_exp,cor(MFD,total_n_fl,use="complete.obs"))
with(data_exp,cor(LFD,total_n_fl,use="complete.obs"))
```


# New appendix: Distances crossings

Read data from csv files:

```{r}
distances_crossings<-read_csv(file="data/clean/distances_crossings.csv")
distances_crossings
```

Histograms:

```{r}
new_new_appendix<-grid.arrange(
  ggplot(distances_crossings,aes(x=dist))+
    geom_histogram(color="black",fill="grey",bins=15)+
    my_theme()+xlab("Distance (m)")+ylab("Number of crossings")+ggtitle("A)"),
  ggplot(data_exp%>%mutate(temp_diff=abs(temp_dam-temp_sire))%>%
           distinct(dam_sire,temp_diff),
         aes(x=temp_diff))+geom_histogram(color="black",fill="grey",bins=15)+
    my_theme()+
    xlab("Absolute temperature difference (ºC)")+
    ylab("Number of crossings")+ggtitle("B)"),
  ncol=2)
ggsave(file="output/figures/new_new_appendix.tiff", plot=new_new_appendix,
       width=12,height=5,dpi=300,compression="lzw")
```

Ranges, means and SDs (for main text):

```{r}
range(distances_crossings$dist)
mean(distances_crossings$dist)
sd(distances_crossings$dist)
```
```{r}
range((data_exp%>%mutate(temp_diff=abs(temp_dam-temp_sire))%>%
         distinct(dam_sire,temp_diff))$temp_diff)
mean((data_exp%>%mutate(temp_diff=abs(temp_dam-temp_sire))%>%
         distinct(dam_sire,temp_diff))$temp_diff)
sd((data_exp%>%mutate(temp_diff=abs(temp_dam-temp_sire))%>%
         distinct(dam_sire,temp_diff))$temp_diff)
```

# Rerun analyses with block

## Read data from csv file

```{r}
data_exp_blocks<-read_csv(file="data/clean/data_exp_blocks.csv")%>%
  mutate(across(c(unique_id, treat,sire,dam,dam_sire,block),as.factor),
         across(c(total_n_fl,FFD,LFD,MFD),as.integer),
         across(c(temp_dam,temp_sire,mean_sh_growth,mean_sh_h_date1,
                  mean_sh_h_date2),as.double))
```

## Additive genetic variation in flowering time

### Models

```{r}
model1_FFD_uh_block<-lmer(FFD~1+(1|sire)+(1|dam)+(1|sire:dam)+(1|block),
                 subset(data_exp_blocks,treat=="unheated"))
model1_FFD_h_block<-lmer(FFD~1+(1|sire)+(1|dam)+(1|sire:dam)+(1|block),
                 subset(data_exp_blocks,treat=="heatmat"))
model1_LFD_uh_block<-lmer(LFD~1+(1|sire)+(1|dam)+(1|sire:dam)+(1|block),
                 subset(data_exp_blocks,treat=="unheated"))
model1_LFD_h_block<-lmer(LFD~1+(1|sire)+(1|dam)+(1|sire:dam)+(1|block),
                 subset(data_exp_blocks,treat=="heatmat"))
model1_MFD_uh_block<-lmer(MFD~1+(1|sire)+(1|dam)+(1|sire:dam)+(1|block),
                 subset(data_exp_blocks,treat=="unheated"))
model1_MFD_h_block<-lmer(MFD~1+(1|sire)+(1|dam)+(1|sire:dam)+(1|block),
                 subset(data_exp_blocks,treat=="heatmat"))
summary(model1_FFD_uh_block)
summary(model1_FFD_h_block)
summary(model1_LFD_uh_block)
summary(model1_LFD_h_block)
summary(model1_MFD_uh_block)
summary(model1_MFD_h_block)
```

### Proportions of variance

```{r}
Variance_FFD_uh_block<-as.data.frame(VarCorr(model1_FFD_uh_block))[,c(1,4)]
Variance_LFD_uh_block<-as.data.frame(VarCorr(model1_LFD_uh_block))[,c(1,4)]
Variance_MFD_uh_block<-as.data.frame(VarCorr(model1_MFD_uh_block))[,c(1,4)]
Variance_FFD_h_block<-as.data.frame(VarCorr(model1_FFD_h_block))[,c(1,4)]
Variance_LFD_h_block<-as.data.frame(VarCorr(model1_LFD_h_block))[,c(1,4)]
Variance_MFD_h_block<-as.data.frame(VarCorr(model1_MFD_h_block))[,c(1,4)]
# Intra-class correlations
PropVar_FFD_uh_block <- Variance_FFD_uh_block%>%
  mutate(propvar=vcov/sum(vcov))%>%mutate(variable="FFD",treat="unheated")
PropVar_LFD_uh_block <- Variance_LFD_uh_block%>%
  mutate(propvar=vcov/sum(vcov))%>%mutate(variable="LFD",treat="unheated")
PropVar_MFD_uh_block <- Variance_MFD_uh_block%>%
  mutate(propvar=vcov/sum(vcov))%>%mutate(variable="MFD",treat="unheated")
PropVar_FFD_h_block <- Variance_FFD_h_block%>%
  mutate(propvar=vcov/sum(vcov))%>%mutate(variable="FFD",treat="heated")
PropVar_LFD_h_block <- Variance_LFD_h_block%>%
  mutate(propvar=vcov/sum(vcov))%>%mutate(variable="LFD",treat="heated")
PropVar_MFD_h_block <- Variance_MFD_h_block%>%
  mutate(propvar=vcov/sum(vcov))%>%mutate(variable="MFD",treat="heated")
Props_var_block<-rbind(PropVar_FFD_uh_block,PropVar_LFD_uh_block,
                       PropVar_MFD_uh_block,PropVar_FFD_h_block,
                       PropVar_LFD_h_block,PropVar_MFD_h_block)
```

```{r}
Props_var_block
```

### Heritability and maternal effects

```{r}
# h^2 (paternal effects)
her_FFD_uh_block<-as.numeric(4*subset(PropVar_FFD_uh_block,grp=="sire")[3])
her_FFD_h_block<-as.numeric(4*subset(PropVar_FFD_h_block,grp=="sire")[3])
her_LFD_uh_block<-as.numeric(4*subset(PropVar_LFD_uh_block,grp=="sire")[3])
her_LFD_h_block<-as.numeric(4*subset(PropVar_LFD_h_block,grp=="sire")[3])
her_MFD_uh_block<-as.numeric(4*subset(PropVar_MFD_uh_block,grp=="sire")[3])
her_MFD_h_block<-as.numeric(4*subset(PropVar_MFD_h_block,grp=="sire")[3])
# Because the additive genetic variance, VA,
# is expected to be four times the among pollen‐donor variance 
# (Falconer & Mackay, 1996; Lynch & Walsh, 1998)
her_block<-data.frame(value=rbind(her_FFD_uh_block,her_FFD_h_block,
                                  her_LFD_uh_block,her_LFD_h_block,
                                  her_MFD_uh_block,her_MFD_h_block))%>%
  rownames_to_column()%>%
  mutate(variable=c("FFD","FFD","LFD","LFD","MFD","MFD"),
         treat=c("unheated","heated","unheated","heated",
                 "unheated","heated"),
         effect="Heritability")
```

```{r}
# maternal effects

# Maternal - paternal
# ----------------------
# summed effects

# Because the pollen‐recipient variance component contains a combination of
# genetic and environmental effects, we subtracted the additive genetic 
# (pollen donor) component from the pollen-recipient variance component
# before dividing the resulting estimate by VP to estimate 
# m2 (m2 = (Vpollen recipient − Vpollen donor)/VP).

mat_FFD_uh_block<-(Variance_FFD_uh_block[2,2]-Variance_FFD_uh_block[3,2])/
  as.numeric(Variance_FFD_uh_block%>%summarise(sum(vcov)))
mat_FFD_h_block<-(Variance_FFD_h_block[2,2]-Variance_FFD_h_block[3,2])/
  as.numeric(Variance_FFD_h_block%>%summarise(sum(vcov)))
mat_LFD_uh_block<-(Variance_LFD_uh_block[2,2]-Variance_LFD_uh_block[3,2])/
  as.numeric(Variance_LFD_uh_block%>%summarise(sum(vcov)))
mat_LFD_h_block<-(Variance_LFD_h_block[2,2]-Variance_LFD_h_block[3,2])/
  as.numeric(Variance_LFD_h_block%>%summarise(sum(vcov)))
mat_MFD_uh_block<-(Variance_MFD_uh_block[2,2]-Variance_MFD_uh_block[3,2])/
  as.numeric(Variance_MFD_uh_block%>%summarise(sum(vcov)))
mat_MFD_h_block<-(Variance_MFD_h_block[2,2]-Variance_MFD_h_block[3,2])/
  as.numeric(Variance_MFD_h_block%>%summarise(sum(vcov)))
mat_block<-data.frame(value=rbind(mat_FFD_uh_block,mat_FFD_h_block,
                                  mat_LFD_uh_block,mat_LFD_h_block,
                                  mat_MFD_uh_block,mat_MFD_h_block))%>%
  rownames_to_column()%>%
  mutate(variable=c("FFD","FFD","LFD","LFD","MFD","MFD"),
         treat=c("unheated","heated","unheated","heated",
                 "unheated","heated"),
         effect="Maternal effects")
```

```{r}
her_mat_block<-rbind(her_block,mat_block)
her_mat_block
```

### Appendix : LRTs for variance components

```{r}
ranova(model1_FFD_uh_block)
ranova(model1_FFD_h_block)
ranova(model1_LFD_uh_block)
ranova(model1_LFD_h_block)
ranova(model1_MFD_uh_block)
ranova(model1_MFD_h_block)
```

## Additive genetic variation in flowering time plasticity

### Models

```{r}
model2_FFD_block<-lmer(FFD~treat+
                    (1|sire)+(1|dam)+
                    (1|treat:sire)+
                    #(1|treat:dam)+ # Removed to avoid singular fit
                    (1|sire:dam)+(1|block),
                  data_exp_blocks)
model2_LFD_block<-lmer(LFD~treat+
                    (1|sire)+(1|dam)+
                    #(1|treat:sire)+ # Removed to avoid singular fit
                    (1|treat:dam)+
                    (1|sire:dam)+(1|block),
                  data_exp_blocks)
model2_MFD_block<-lmer(MFD~treat+
                    (1|sire)+(1|dam)+
                    (1|treat:sire)+
                    (1|treat:dam)+ # No need to remove to aovid singular fit!
                    (1|sire:dam)+(1|block),
                  data_exp_blocks)
summary(model2_FFD_block)
summary(model2_LFD_block)
summary(model2_MFD_block)
```

### Proportions of variance

```{r}
# extract variance components
Variance_FFD_treat_block <- as.data.frame(VarCorr(model2_FFD_block))[,c(1,4)]  
Variance_LFD_treat_block <- as.data.frame(VarCorr(model2_LFD_block))[,c(1,4)]  
Variance_MFD_treat_block <- as.data.frame(VarCorr(model2_MFD_block))[,c(1,4)]  
# Intra-class correlation
PropVar_FFD_treat_block <- Variance_FFD_treat_block%>%
  mutate(propvar=vcov/sum(vcov))%>%mutate(variable="FFD")
PropVar_LFD_treat_block <- Variance_LFD_treat_block%>%
  mutate(propvar=vcov/sum(vcov))%>%mutate(variable="LFD")
PropVar_MFD_treat_block <- Variance_MFD_treat_block%>%
  mutate(propvar=vcov/sum(vcov))%>%mutate(variable="MFD")
# Proportional variance
Props_var_treat_block<-rbind(PropVar_FFD_treat_block,PropVar_LFD_treat_block,
                       PropVar_MFD_treat_block)
```

```{r}
Props_var_treat_block
```

### Heritability and maternal effects

```{r}
# h^2 (paternal effects)
her_FFD_treat_block<-4*subset(PropVar_FFD_treat_block,grp=="treat:sire")[3]
her_LFD_treat_block<-0
her_MFD_treat_block<-4*subset(PropVar_MFD_treat_block,grp=="treat:sire")[3]
# Because the additive genetic variance, VA,
# is expected to be four times the among pollen‐donor variance 
# (Falconer & Mackay, 1996; Lynch & Walsh, 1998)
her_treat_block<-data.frame(value=rbind(her_FFD_treat_block,
                                  her_LFD_treat_block,her_MFD_treat_block))%>%
  mutate(variable=c("FFD","LFD","MFD"),
         effect="Heritability")%>%
  rename(value=propvar)
```

```{r}
# maternal effects

# Maternal - paternal
# ----------------------
# summed effects

# Because the pollen‐recipient variance component contains a combination of
# genetic and environmental effects, we subtracted the additive genetic 
# (pollen donor) component from the pollen-recipient variance component
# before dividing the resulting estimate by VP to estimate 
# m2 (m2 = (Vpollen recipient − Vpollen donor)/VP).

mat_FFD_treat_block<-(0-
                  subset(Variance_FFD_treat_block,grp=="treat:sire")[2])/
  Variance_FFD_treat_block%>%summarise(sum(vcov))
mat_LFD_treat_block<-(subset(Variance_LFD_treat_block,grp=="treat:dam")[2]-
                  0)/
  Variance_LFD_treat_block%>%summarise(sum(vcov))
mat_MFD_treat_block<-(subset(Variance_MFD_treat_block,grp=="treat:dam")[2]-
                  subset(Variance_MFD_treat_block,grp=="treat:sire")[2])/
  Variance_MFD_treat_block%>%summarise(sum(vcov))
mat_treat_block<-data.frame(value=rbind(mat_FFD_treat_block,
                                  mat_LFD_treat_block,mat_MFD_treat_block))%>%
  mutate(variable=c("FFD","LFD","MFD"),
         effect="Maternal effects")%>%
  rename(value=vcov)
```

```{r}
her_mat_treat_block<-rbind(her_treat_block,mat_treat_block)
her_mat_treat_block
```

### Appendix X: LRTs for variance components

```{r}
ranova(model2_FFD_block)
ranova(model2_LFD_block)
ranova(model2_MFD_block)
```

## Table 1

```{r}
her_mat_block<-her_mat_block%>%
  mutate(variable=factor(variable,levels=c("FFD","MFD","LFD")),
         treat=factor(treat,levels=c("unheated","heated")))%>%
  dplyr::select(-rowname)%>%
  pivot_wider(names_from="effect",values_from="value")
her_mat_block
her_mat_treat_block<-her_mat_treat_block%>%
  pivot_wider(names_from="effect",values_from="value")
her_mat_treat_block
```

## Hypothesis 1

Temperature-dependent selection has resulted in small-scale genetic differentiation, and additive genetic variation in phenology is correlated with soil temperature in a counter-gradient fashion. At a given temperature, individuals with parents from warmer soils are predicted to start vegetative growth and flower later than plants with parents from colder soils.

### Models temperatures sire and dam

```{r}
model3_FFD_uh_block<-lmer(FFD~temp_sire+temp_dam+
                      (1|sire)+(1|dam)+(1|sire:dam),
                      #+(1|block), # Removed to avoid singular fit
                 subset(data_exp_blocks,treat=="unheated"&!is.na(FFD)))
model3_FFD_h_block<-lmer(FFD~temp_sire+temp_dam+
                     (1|sire)+(1|dam)+(1|sire:dam)+(1|block),
                 subset(data_exp_blocks,treat=="heatmat"&!is.na(FFD)))
model3_LFD_uh_block<-lmer(LFD~temp_sire+temp_dam+
                      (1|sire)+(1|dam)+(1|sire:dam),
                      #+(1|block), # Removed to avoid singular fit
                 subset(data_exp_blocks,treat=="unheated"&!is.na(LFD)))
model3_LFD_h_block<-lmer(LFD~temp_sire+temp_dam+
                     #(1|sire)+ # Removed to avoid singular fit
                       (1|dam)+(1|sire:dam)+(1|block),
                 subset(data_exp_blocks,treat=="heatmat"&!is.na(LFD)))
model3_MFD_uh_block<-lmer(MFD~temp_sire+temp_dam+
                      (1|sire)+(1|dam)+(1|sire:dam),
                      #+(1|block), # Removed to avoid singular fit
                 subset(data_exp_blocks,treat=="unheated"&!is.na(MFD)))
model3_MFD_h_block<-lmer(MFD~temp_sire+temp_dam+
                     (1|sire)+ # No need to remove to avoid singular fit!
                     (1|dam)+(1|sire:dam)+(1|block),
                 subset(data_exp_blocks,treat=="heatmat"&!is.na(MFD)))
summary(model3_FFD_uh_block)
summary(model3_FFD_h_block)
summary(model3_LFD_uh_block)
summary(model3_LFD_h_block)
summary(model3_MFD_uh_block)
summary(model3_MFD_h_block)
```

#### Table 2

```{r}
tab_model(model3_FFD_uh_block,model3_MFD_uh_block,model3_LFD_uh_block,
          transform=NULL,show.intercept=F,show.ci=F,show.se=T,show.stat=T,
          show.r2=T,show.icc=F,show.re.var=F,show.ngroups=F,digits=3,
          dv.labels=c("FFD","MFD","LFD"),
          file="output/tables/Table2_uh_block.html")
tab_model(model3_FFD_h_block,model3_MFD_h_block,model3_LFD_h_block,
          transform=NULL,show.intercept=F,show.ci=F,show.se=T,show.stat=T,
          show.r2=T,show.icc=F,show.re.var=F,show.ngroups=F,digits=3,
          dv.labels=c("FFD","MFD","LFD"),
          file="output/tables/Table2_h_block.html")
```

## Hypothesis 2

Genetic differentiation related to soil temperature involves also plasticity of development rate. Differences in phenology between high and low temperatures are predicted to be smaller for plants with parents from warm soils than for plants from colder soils.

### Models temperatures sire and dam

```{r}
model4_FFD_block<-lmer(FFD~treat*(temp_sire+temp_dam)+
                    (1|sire)+(1|dam)+
                    (1|sire:dam)+(1|block),
                  data_exp_blocks)
model4_LFD_block<-lmer(LFD~treat*(temp_sire+temp_dam)+
                    (1|sire)+(1|dam)+
                    (1|sire:dam)+(1|block),
                  data_exp_blocks)
model4_MFD_block<-lmer(MFD~treat*(temp_sire+temp_dam)+
                    (1|sire)+(1|dam)+
                    (1|sire:dam)+(1|block),
                  subset(data_exp_blocks,!is.na(MFD)))
summary(model4_FFD_block)
summary(model4_LFD_block)
summary(model4_MFD_block)
# All singular fit!
```

# Session info

```{r}
sessionInfo()
```
